Create an object summarizing both continuous and categorical variables

Create an object summarizing all baseline variables (both continuous and categorical) optionally stratifying by one or more startifying variables and performing statistical tests. The object gives a table that is easy to use in medical research papers.

CreateTableOne(
  vars,
  strata,
  data,
  factorVars,
  includeNA = FALSE,
  test = TRUE,
  testApprox = chisq.test,
  argsApprox = list(correct = TRUE),
  testExact = fisher.test,
  argsExact = list(workspace = 2 * 10^5),
  testNormal = oneway.test,
  argsNormal = list(var.equal = TRUE),
  testNonNormal = kruskal.test,
  argsNonNormal = list(NULL),
  smd = TRUE,
  addOverall = FALSE
)

Arguments

vars

Variables to be summarized given as a character vector. Factors are handled as categorical variables, whereas numeric variables are handled as continuous variables. If empty, all variables in the data frame specified in the data argument are used.

strata

Stratifying (grouping) variable name(s) given as a character vector. If omitted, the overall results are returned.

data

A data frame in which these variables exist. All variables (both vars and strata) must be in this data frame.

factorVars

Numerically coded variables that should be handled as categorical variables given as a character vector. Do not include factors, unless you need to relevel them by removing empty levels. If omitted, only factors are considered categorical variables. The variables specified here must also be specified in the vars argument.

includeNA

If TRUE, NA is handled as a regular factor level rather than missing. NA is shown as the last factor level in the table. Only effective for categorical variables.

test

If TRUE, as in the default and there are more than two groups, groupwise comparisons are performed.

testApprox

A function used to perform the large sample approximation based tests. The default is chisq.test. This is not recommended when some of the cell have small counts like fewer than 5.

argsApprox

A named list of arguments passed to the function specified in testApprox. The default is list(correct = TRUE), which turns on the continuity correction for chisq.test.

testExact

A function used to perform the exact tests. The default is fisher.test. If the cells have large numbers, it will fail because of memory limitation. In this situation, the large sample approximation based should suffice.

argsExact

A named list of arguments passed to the function specified in testExact. The default is list(workspace = 2*10^5), which specifies the memory space allocated for fisher.test.

testNormal

A function used to perform the normal assumption based tests. The default is oneway.test. This is equivalent of the t-test when there are only two groups.

argsNormal

A named list of arguments passed to the function specified in testNormal. The default is list(var.equal = TRUE), which makes it the ordinary ANOVA that assumes equal variance across groups.

testNonNormal

A function used to perform the nonparametric tests. The default is kruskal.test (Kruskal-Wallis Rank Sum Test). This is equivalent of the wilcox.test (Mann-Whitney U test) when there are only two groups.

argsNonNormal

A named list of arguments passed to the function specified in testNonNormal. The default is list(NULL), which is just a placeholder.

smd

If TRUE, as in the default and there are more than two groups, standardized mean differences for all pairwise comparisons are calculated.

addOverall

(optional, only used if strata are supplied) Adds an overall column to the table. Smd and p-value calculations are performed using only the stratifed clolumns.

Value

An object of class TableOne, which is a list of three objects.

ContTable

object of class ContTable, containing continuous variables only

CatTable

object of class CatTable, containing categorical variables only

MetaData

list of metadata regarding variables

Details

The definitions of the standardized mean difference (SMD) are available in Flury et al 1986 for the univariate case and the multivariate case (essentially the square root of the Mahalanobis distance). Extension to binary variables is discussed in Austin 2009 and extension to multinomival variables is suggested in Yang et al 2012. This multinomial extesion treats a single multinomial variable as multiple non-redundant dichotomous variables and use the Mahalanobis distance. The off diagonal elements of the covariance matrix on page 3 have an error, and need negation. In weighted data, the same definitions can be used except that the mean and standard deviation estimates are weighted estimates (Li et al 2013 and Austin et al 2015). In tableone, all weighted estimates are calculated by weighted estimation functions in the survey package.

References

Flury, BK. and Riedwyl, H. (1986). Standard distance in univariate and multivariate analysis. The American Statistician, 40, 249-251.

Austin, PC. (2009). Using the Standardized Difference to Compare the Prevalence of a Binary Variable Between Two Groups in Observational Research. Communications in Statistics - Simulation and Computation, 38, 1228-1234.

Yang, D. and Dalton, JE. (2012). A unified approach to measuring the effect size between two groups using SAS. SAS Global Forum 2012, Paper 335-2012.

Li, L. and Greene, T. (2013). A weighting analogue to pair matching in propensity score analysis. International Journal of Biostatistics, 9, 215-234.

Austin, PC. and Stuart, EA. (2015). Moving towards best practice when using inverse probability of treatment weighting (IPTW) using the propensity score to estimate causal treatment effects in observational studies. Statistics in Medicine, Online on August 3, 2015.

See also

Author

Kazuki Yoshida, Justin Bohn, Alexander Bartel

Examples

## Load
library(tableone)

## Load Mayo Clinic Primary Biliary Cirrhosis Data
library(survival)
data(pbc)
## Check variables
head(pbc)
#>   id time status trt      age sex ascites hepato spiders edema bili chol
#> 1  1  400      2   1 58.76523   f       1      1       1   1.0 14.5  261
#> 2  2 4500      0   1 56.44627   f       0      1       1   0.0  1.1  302
#> 3  3 1012      2   1 70.07255   m       0      0       0   0.5  1.4  176
#> 4  4 1925      2   1 54.74059   f       0      1       1   0.5  1.8  244
#> 5  5 1504      1   2 38.10541   f       0      1       1   0.0  3.4  279
#> 6  6 2503      2   2 66.25873   f       0      1       0   0.0  0.8  248
#>   albumin copper alk.phos    ast trig platelet protime stage
#> 1    2.60    156   1718.0 137.95  172      190    12.2     4
#> 2    4.14     54   7394.8 113.52   88      221    10.6     3
#> 3    3.48    210    516.0  96.10   55      151    12.0     4
#> 4    2.54     64   6121.8  60.63   92      183    10.3     4
#> 5    3.53    143    671.0 113.15   72      136    10.9     3
#> 6    3.98     50    944.0  93.00   63       NA    11.0     3

## Make categorical variables factors
varsToFactor <- c("status","trt","ascites","hepato","spiders","edema","stage")
pbc[varsToFactor] <- lapply(pbc[varsToFactor], factor)

## Create a variable list
dput(names(pbc))
#> c("id", "time", "status", "trt", "age", "sex", "ascites", "hepato", 
#> "spiders", "edema", "bili", "chol", "albumin", "copper", "alk.phos", 
#> "ast", "trig", "platelet", "protime", "stage")
vars <- c("time","status","age","sex","ascites","hepato",
          "spiders","edema","bili","chol","albumin",
          "copper","alk.phos","ast","trig","platelet",
          "protime","stage")

## Create Table 1 stratified by trt
tableOne <- CreateTableOne(vars = vars, strata = c("trt"), data = pbc)

## Just typing the object name will invoke the print.TableOne method
tableOne
#>                       Stratified by trt
#>                        1                 2                 p      test
#>   n                        158               154                      
#>   time (mean (SD))     2015.62 (1094.12) 1996.86 (1155.93)  0.883     
#>   status (%)                                                0.894     
#>      0                      83 (52.5)         85 (55.2)               
#>      1                      10 ( 6.3)          9 ( 5.8)               
#>      2                      65 (41.1)         60 (39.0)               
#>   age (mean (SD))        51.42 (11.01)     48.58 (9.96)     0.018     
#>   sex = f (%)              137 (86.7)        139 (90.3)     0.421     
#>   ascites = 1 (%)           14 ( 8.9)         10 ( 6.5)     0.567     
#>   hepato = 1 (%)            73 (46.2)         87 (56.5)     0.088     
#>   spiders = 1 (%)           45 (28.5)         45 (29.2)     0.985     
#>   edema (%)                                                 0.877     
#>      0                     132 (83.5)        131 (85.1)               
#>      0.5                    16 (10.1)         13 ( 8.4)               
#>      1                      10 ( 6.3)         10 ( 6.5)               
#>   bili (mean (SD))        2.87 (3.63)       3.65 (5.28)     0.131     
#>   chol (mean (SD))      365.01 (209.54)   373.88 (252.48)   0.748     
#>   albumin (mean (SD))     3.52 (0.44)       3.52 (0.40)     0.874     
#>   copper (mean (SD))     97.64 (90.59)     97.65 (80.49)    0.999     
#>   alk.phos (mean (SD)) 2021.30 (2183.44) 1943.01 (2101.69)  0.747     
#>   ast (mean (SD))       120.21 (54.52)    124.97 (58.93)    0.460     
#>   trig (mean (SD))      124.14 (71.54)    125.25 (58.52)    0.886     
#>   platelet (mean (SD))  258.75 (100.32)   265.20 (90.73)    0.555     
#>   protime (mean (SD))    10.65 (0.85)      10.80 (1.14)     0.197     
#>   stage (%)                                                 0.201     
#>      1                      12 ( 7.6)          4 ( 2.6)               
#>      2                      35 (22.2)         32 (20.8)               
#>      3                      56 (35.4)         64 (41.6)               
#>      4                      55 (34.8)         54 (35.1)               

## Specifying nonnormal variables will show the variables appropriately,
## and show nonparametric test p-values. Specify variables in the exact
## argument to obtain the exact test p-values. cramVars can be used to
## show both levels for a 2-level categorical variables.
print(tableOne, nonnormal = c("bili","chol","copper","alk.phos","trig"),
      exact = c("status","stage"), cramVars = "hepato", smd = TRUE)
#>                          Stratified by trt
#>                           1                         2                        
#>   n                           158                       154                  
#>   time (mean (SD))        2015.62 (1094.12)         1996.86 (1155.93)        
#>   status (%)                                                                 
#>      0                         83 (52.5)                 85 (55.2)           
#>      1                         10 ( 6.3)                  9 ( 5.8)           
#>      2                         65 (41.1)                 60 (39.0)           
#>   age (mean (SD))           51.42 (11.01)             48.58 (9.96)           
#>   sex = f (%)                 137 (86.7)                139 (90.3)           
#>   ascites = 1 (%)              14 ( 8.9)                 10 ( 6.5)           
#>   hepato = 0/1 (%)          85/73 (53.8/46.2)         67/87 (43.5/56.5)      
#>   spiders = 1 (%)              45 (28.5)                 45 (29.2)           
#>   edema (%)                                                                  
#>      0                        132 (83.5)                131 (85.1)           
#>      0.5                       16 (10.1)                 13 ( 8.4)           
#>      1                         10 ( 6.3)                 10 ( 6.5)           
#>   bili (median [IQR])        1.40 [0.80, 3.20]         1.30 [0.72, 3.60]     
#>   chol (median [IQR])      315.50 [247.75, 417.00]   303.50 [254.25, 377.00] 
#>   albumin (mean (SD))        3.52 (0.44)               3.52 (0.40)           
#>   copper (median [IQR])     73.00 [40.00, 121.00]     73.00 [43.00, 139.00]  
#>   alk.phos (median [IQR]) 1214.50 [840.75, 2028.00] 1283.00 [922.50, 1949.75]
#>   ast (mean (SD))          120.21 (54.52)            124.97 (58.93)          
#>   trig (median [IQR])      106.00 [84.50, 146.00]    113.00 [84.50, 155.00]  
#>   platelet (mean (SD))     258.75 (100.32)           265.20 (90.73)          
#>   protime (mean (SD))       10.65 (0.85)              10.80 (1.14)           
#>   stage (%)                                                                  
#>      1                         12 ( 7.6)                  4 ( 2.6)           
#>      2                         35 (22.2)                 32 (20.8)           
#>      3                         56 (35.4)                 64 (41.6)           
#>      4                         55 (34.8)                 54 (35.1)           
#>                          Stratified by trt
#>                           p      test    SMD   
#>   n                                            
#>   time (mean (SD))         0.883          0.017
#>   status (%)               0.884 exact    0.054
#>      0                                         
#>      1                                         
#>      2                                         
#>   age (mean (SD))          0.018          0.270
#>   sex = f (%)              0.421          0.111
#>   ascites = 1 (%)          0.567          0.089
#>   hepato = 0/1 (%)         0.088          0.207
#>   spiders = 1 (%)          0.985          0.016
#>   edema (%)                0.877          0.058
#>      0                                         
#>      0.5                                       
#>      1                                         
#>   bili (median [IQR])      0.842 nonnorm  0.171
#>   chol (median [IQR])      0.544 nonnorm  0.038
#>   albumin (mean (SD))      0.874          0.018
#>   copper (median [IQR])    0.717 nonnorm <0.001
#>   alk.phos (median [IQR])  0.812 nonnorm  0.037
#>   ast (mean (SD))          0.460          0.084
#>   trig (median [IQR])      0.370 nonnorm  0.017
#>   platelet (mean (SD))     0.555          0.067
#>   protime (mean (SD))      0.197          0.146
#>   stage (%)                0.205 exact    0.246
#>      1                                         
#>      2                                         
#>      3                                         
#>      4                                         

## Use the summary.TableOne method for detailed summary
summary(tableOne)
#> 
#>      ### Summary of continuous variables ###
#> 
#> trt: 1
#>            n miss p.miss mean    sd median   p25  p75   min   max  skew kurt
#> time     158    0    0.0 2016 1e+03   1895 1e+03 2632  41.0  4556  0.41 -0.4
#> age      158    0    0.0   51 1e+01     52 4e+01   59  26.3    78  0.06 -0.5
#> bili     158    0    0.0    3 4e+00      1 8e-01    3   0.3    20  2.67  7.6
#> chol     158   18   11.4  365 2e+02    316 2e+02  417 127.0  1712  3.83 20.2
#> albumin  158    0    0.0    4 4e-01      4 3e+00    4   2.1     5 -0.40  0.3
#> copper   158    1    0.6   98 9e+01     73 4e+01  121   9.0   588  2.50  8.2
#> alk.phos 158    0    0.0 2021 2e+03   1214 8e+02 2028 369.0 11552  2.71  7.4
#> ast      158    0    0.0  120 5e+01    112 8e+01  152  26.4   338  1.09  1.6
#> trig     158   19   12.0  124 7e+01    106 8e+01  146  33.0   598  2.95 14.3
#> platelet 158    2    1.3  259 1e+02    255 2e+02  322  62.0   563  0.50  0.2
#> protime  158    0    0.0   11 9e-01     11 1e+01   11   9.0    14  1.10  1.6
#> ------------------------------------------------------------ 
#> trt: 2
#>            n miss p.miss mean    sd median   p25  p75   min   max skew kurt
#> time     154    0    0.0 1997 1e+03   1811 1e+03 2771  51.0  4523  0.4 -0.7
#> age      154    0    0.0   49 1e+01     48 4e+01   56  30.6    75  0.2 -0.5
#> bili     154    0    0.0    4 5e+00      1 7e-01    4   0.3    28  2.7  7.3
#> chol     154   10    6.5  374 3e+02    304 3e+02  377 120.0  1775  3.1 11.1
#> albumin  154    0    0.0    4 4e-01      4 3e+00    4   2.0     4 -0.8  2.0
#> copper   154    1    0.6   98 8e+01     73 4e+01  139   4.0   558  2.0  6.6
#> alk.phos 154    0    0.0 1943 2e+03   1283 9e+02 1950 289.0 13862  3.3 12.8
#> ast      154    0    0.0  125 6e+01    117 8e+01  152  28.4   457  1.7  6.3
#> trig     154   11    7.1  125 6e+01    113 8e+01  155  44.0   432  1.7  5.5
#> platelet 154    2    1.3  265 9e+01    260 2e+02  322  71.0   487  0.2 -0.3
#> protime  154    0    0.0   11 1e+00     11 1e+01   11   9.2    17  1.9  6.4
#> 
#> p-values
#>             pNormal pNonNormal
#> time     0.88304691 0.82661809
#> age      0.01767247 0.01962155
#> bili     0.13093942 0.84168460
#> chol     0.74799072 0.54433899
#> albumin  0.87388074 0.95045176
#> copper   0.99915849 0.71745444
#> alk.phos 0.74726165 0.81198200
#> ast      0.45969842 0.45892358
#> trig     0.88604213 0.36980434
#> platelet 0.55451136 0.45482564
#> protime  0.19714026 0.58802048
#> 
#> Standardize mean differences
#>                1 vs 2
#> time     0.0166658751
#> age      0.2702619258
#> bili     0.1710905651
#> chol     0.0382210537
#> albumin  0.0180021838
#> copper   0.0001200022
#> alk.phos 0.0365323630
#> ast      0.0837836058
#> trig     0.0170615337
#> platelet 0.0674763888
#> protime  0.1460939117
#> 
#> =======================================================================================
#> 
#>      ### Summary of categorical variables ### 
#> 
#> trt: 1
#>      var   n miss p.miss level freq percent cum.percent
#>   status 158    0    0.0     0   83    52.5        52.5
#>                              1   10     6.3        58.9
#>                              2   65    41.1       100.0
#>                                                        
#>      sex 158    0    0.0     m   21    13.3        13.3
#>                              f  137    86.7       100.0
#>                                                        
#>  ascites 158    0    0.0     0  144    91.1        91.1
#>                              1   14     8.9       100.0
#>                                                        
#>   hepato 158    0    0.0     0   85    53.8        53.8
#>                              1   73    46.2       100.0
#>                                                        
#>  spiders 158    0    0.0     0  113    71.5        71.5
#>                              1   45    28.5       100.0
#>                                                        
#>    edema 158    0    0.0     0  132    83.5        83.5
#>                            0.5   16    10.1        93.7
#>                              1   10     6.3       100.0
#>                                                        
#>    stage 158    0    0.0     1   12     7.6         7.6
#>                              2   35    22.2        29.7
#>                              3   56    35.4        65.2
#>                              4   55    34.8       100.0
#>                                                        
#> ------------------------------------------------------------ 
#> trt: 2
#>      var   n miss p.miss level freq percent cum.percent
#>   status 154    0    0.0     0   85    55.2        55.2
#>                              1    9     5.8        61.0
#>                              2   60    39.0       100.0
#>                                                        
#>      sex 154    0    0.0     m   15     9.7         9.7
#>                              f  139    90.3       100.0
#>                                                        
#>  ascites 154    0    0.0     0  144    93.5        93.5
#>                              1   10     6.5       100.0
#>                                                        
#>   hepato 154    0    0.0     0   67    43.5        43.5
#>                              1   87    56.5       100.0
#>                                                        
#>  spiders 154    0    0.0     0  109    70.8        70.8
#>                              1   45    29.2       100.0
#>                                                        
#>    edema 154    0    0.0     0  131    85.1        85.1
#>                            0.5   13     8.4        93.5
#>                              1   10     6.5       100.0
#>                                                        
#>    stage 154    0    0.0     1    4     2.6         2.6
#>                              2   32    20.8        23.4
#>                              3   64    41.6        64.9
#>                              4   54    35.1       100.0
#>                                                        
#> 
#> p-values
#>            pApprox     pExact
#> status  0.89350975 0.88422188
#> sex     0.42122610 0.37743235
#> ascites 0.56728647 0.52558267
#> hepato  0.08820884 0.07137522
#> spiders 0.98466036 0.90113734
#> edema   0.87681949 0.89370131
#> stage   0.20129629 0.20455558
#> 
#> Standardize mean differences
#>             1 vs 2
#> status  0.05375763
#> sex     0.11141161
#> ascites 0.08900618
#> hepato  0.20699413
#> spiders 0.01632844
#> edema   0.05811659
#> stage   0.24600834

## See the categorical part only using $ operator
tableOne$CatTable
#>                  Stratified by trt
#>                   1           2           p      test
#>   n               158         154                    
#>   status (%)                               0.894     
#>      0             83 (52.5)   85 (55.2)             
#>      1             10 ( 6.3)    9 ( 5.8)             
#>      2             65 (41.1)   60 (39.0)             
#>   sex = f (%)     137 (86.7)  139 (90.3)   0.421     
#>   ascites = 1 (%)  14 ( 8.9)   10 ( 6.5)   0.567     
#>   hepato = 1 (%)   73 (46.2)   87 (56.5)   0.088     
#>   spiders = 1 (%)  45 (28.5)   45 (29.2)   0.985     
#>   edema (%)                                0.877     
#>      0            132 (83.5)  131 (85.1)             
#>      0.5           16 (10.1)   13 ( 8.4)             
#>      1             10 ( 6.3)   10 ( 6.5)             
#>   stage (%)                                0.201     
#>      1             12 ( 7.6)    4 ( 2.6)             
#>      2             35 (22.2)   32 (20.8)             
#>      3             56 (35.4)   64 (41.6)             
#>      4             55 (34.8)   54 (35.1)             
summary(tableOne$CatTable)
#> trt: 1
#>      var   n miss p.miss level freq percent cum.percent
#>   status 158    0    0.0     0   83    52.5        52.5
#>                              1   10     6.3        58.9
#>                              2   65    41.1       100.0
#>                                                        
#>      sex 158    0    0.0     m   21    13.3        13.3
#>                              f  137    86.7       100.0
#>                                                        
#>  ascites 158    0    0.0     0  144    91.1        91.1
#>                              1   14     8.9       100.0
#>                                                        
#>   hepato 158    0    0.0     0   85    53.8        53.8
#>                              1   73    46.2       100.0
#>                                                        
#>  spiders 158    0    0.0     0  113    71.5        71.5
#>                              1   45    28.5       100.0
#>                                                        
#>    edema 158    0    0.0     0  132    83.5        83.5
#>                            0.5   16    10.1        93.7
#>                              1   10     6.3       100.0
#>                                                        
#>    stage 158    0    0.0     1   12     7.6         7.6
#>                              2   35    22.2        29.7
#>                              3   56    35.4        65.2
#>                              4   55    34.8       100.0
#>                                                        
#> ------------------------------------------------------------ 
#> trt: 2
#>      var   n miss p.miss level freq percent cum.percent
#>   status 154    0    0.0     0   85    55.2        55.2
#>                              1    9     5.8        61.0
#>                              2   60    39.0       100.0
#>                                                        
#>      sex 154    0    0.0     m   15     9.7         9.7
#>                              f  139    90.3       100.0
#>                                                        
#>  ascites 154    0    0.0     0  144    93.5        93.5
#>                              1   10     6.5       100.0
#>                                                        
#>   hepato 154    0    0.0     0   67    43.5        43.5
#>                              1   87    56.5       100.0
#>                                                        
#>  spiders 154    0    0.0     0  109    70.8        70.8
#>                              1   45    29.2       100.0
#>                                                        
#>    edema 154    0    0.0     0  131    85.1        85.1
#>                            0.5   13     8.4        93.5
#>                              1   10     6.5       100.0
#>                                                        
#>    stage 154    0    0.0     1    4     2.6         2.6
#>                              2   32    20.8        23.4
#>                              3   64    41.6        64.9
#>                              4   54    35.1       100.0
#>                                                        
#> 
#> p-values
#>            pApprox     pExact
#> status  0.89350975 0.88422188
#> sex     0.42122610 0.37743235
#> ascites 0.56728647 0.52558267
#> hepato  0.08820884 0.07137522
#> spiders 0.98466036 0.90113734
#> edema   0.87681949 0.89370131
#> stage   0.20129629 0.20455558
#> 
#> Standardize mean differences
#>             1 vs 2
#> status  0.05375763
#> sex     0.11141161
#> ascites 0.08900618
#> hepato  0.20699413
#> spiders 0.01632844
#> edema   0.05811659
#> stage   0.24600834

## See the continuous part only using $ operator
tableOne$ContTable
#>                       Stratified by trt
#>                        1                 2                 p      test
#>   n                    158               154                          
#>   time (mean (SD))     2015.62 (1094.12) 1996.86 (1155.93)  0.883     
#>   age (mean (SD))        51.42 (11.01)     48.58 (9.96)     0.018     
#>   bili (mean (SD))        2.87 (3.63)       3.65 (5.28)     0.131     
#>   chol (mean (SD))      365.01 (209.54)   373.88 (252.48)   0.748     
#>   albumin (mean (SD))     3.52 (0.44)       3.52 (0.40)     0.874     
#>   copper (mean (SD))     97.64 (90.59)     97.65 (80.49)    0.999     
#>   alk.phos (mean (SD)) 2021.30 (2183.44) 1943.01 (2101.69)  0.747     
#>   ast (mean (SD))       120.21 (54.52)    124.97 (58.93)    0.460     
#>   trig (mean (SD))      124.14 (71.54)    125.25 (58.52)    0.886     
#>   platelet (mean (SD))  258.75 (100.32)   265.20 (90.73)    0.555     
#>   protime (mean (SD))    10.65 (0.85)      10.80 (1.14)     0.197     
summary(tableOne$ContTable)
#> trt: 1
#>            n miss p.miss   mean      sd median    p25    p75   min     max
#> time     158    0   0.00 2015.6 1094.12 1895.0 1231.0 2632.5  41.0  4556.0
#> age      158    0   0.00   51.4   11.01   51.9   43.0   58.9  26.3    78.4
#> bili     158    0   0.00    2.9    3.63    1.4    0.8    3.2   0.3    20.0
#> chol     158   18  11.39  365.0  209.54  315.5  247.8  417.0 127.0  1712.0
#> albumin  158    0   0.00    3.5    0.44    3.6    3.2    3.8   2.1     4.6
#> copper   158    1   0.63   97.6   90.59   73.0   40.0  121.0   9.0   588.0
#> alk.phos 158    0   0.00 2021.3 2183.44 1214.5  840.8 2028.0 369.0 11552.0
#> ast      158    0   0.00  120.2   54.52  111.6   76.7  151.5  26.4   338.0
#> trig     158   19  12.03  124.1   71.54  106.0   84.5  146.0  33.0   598.0
#> platelet 158    2   1.27  258.8  100.32  255.0  189.5  322.0  62.0   563.0
#> protime  158    0   0.00   10.7    0.85   10.6   10.0   11.0   9.0    14.1
#>            skew  kurt
#> time      0.412 -0.44
#> age       0.056 -0.53
#> bili      2.675  7.61
#> chol      3.831 20.21
#> albumin  -0.395  0.30
#> copper    2.503  8.21
#> alk.phos  2.707  7.36
#> ast       1.095  1.60
#> trig      2.947 14.29
#> platelet  0.498  0.25
#> protime   1.103  1.62
#> ------------------------------------------------------------ 
#> trt: 2
#>            n miss p.miss   mean     sd median     p25    p75   min     max
#> time     154    0   0.00 1996.9 1155.9 1811.0 1153.00 2771.2  51.0  4523.0
#> age      154    0   0.00   48.6   10.0   48.1   41.43   55.8  30.6    74.5
#> bili     154    0   0.00    3.6    5.3    1.3    0.72    3.6   0.3    28.0
#> chol     154   10   6.49  373.9  252.5  303.5  254.25  377.0 120.0  1775.0
#> albumin  154    0   0.00    3.5    0.4    3.5    3.34    3.8   2.0     4.4
#> copper   154    1   0.65   97.7   80.5   73.0   43.00  139.0   4.0   558.0
#> alk.phos 154    0   0.00 1943.0 2101.7 1283.0  922.50 1949.8 289.0 13862.4
#> ast      154    0   0.00  125.0   58.9  117.4   83.78  151.9  28.4   457.2
#> trig     154   11   7.14  125.3   58.5  113.0   84.50  155.0  44.0   432.0
#> platelet 154    2   1.30  265.2   90.7  259.5  206.75  322.5  71.0   487.0
#> protime  154    0   0.00   10.8    1.1   10.6   10.00   11.4   9.2    17.1
#>           skew  kurt
#> time      0.35 -0.72
#> age       0.24 -0.53
#> bili      2.67  7.27
#> chol      3.12 11.14
#> albumin  -0.85  1.96
#> copper    2.01  6.65
#> alk.phos  3.35 12.81
#> ast       1.73  6.31
#> trig      1.72  5.51
#> platelet  0.23 -0.35
#> protime   1.87  6.39
#> 
#> p-values
#>             pNormal pNonNormal
#> time     0.88304691 0.82661809
#> age      0.01767247 0.01962155
#> bili     0.13093942 0.84168460
#> chol     0.74799072 0.54433899
#> albumin  0.87388074 0.95045176
#> copper   0.99915849 0.71745444
#> alk.phos 0.74726165 0.81198200
#> ast      0.45969842 0.45892358
#> trig     0.88604213 0.36980434
#> platelet 0.55451136 0.45482564
#> protime  0.19714026 0.58802048
#> 
#> Standardize mean differences
#>                1 vs 2
#> time     0.0166658751
#> age      0.2702619258
#> bili     0.1710905651
#> chol     0.0382210537
#> albumin  0.0180021838
#> copper   0.0001200022
#> alk.phos 0.0365323630
#> ast      0.0837836058
#> trig     0.0170615337
#> platelet 0.0674763888
#> protime  0.1460939117

## If your work flow includes copying to Excel and Word when writing manuscripts,
## you may benefit from the quote argument. This will quote everything so that
## Excel does not mess up the cells.
print(tableOne, nonnormal = c("bili","chol","copper","alk.phos","trig"),
      exact = c("status","stage"), quote = TRUE)
#>                            "Stratified by trt"
#>  ""                         "1"                        
#>   "n"                       "    158"                  
#>   "time (mean (SD))"        "2015.62 (1094.12)"        
#>   "status (%)"              "    "                     
#>   "   0"                    "     83 (52.5) "          
#>   "   1"                    "     10 ( 6.3) "          
#>   "   2"                    "     65 (41.1) "          
#>   "age (mean (SD))"         "  51.42 (11.01)"          
#>   "sex = f (%)"             "    137 (86.7) "          
#>   "ascites = 1 (%)"         "     14 ( 8.9) "          
#>   "hepato = 1 (%)"          "     73 (46.2) "          
#>   "spiders = 1 (%)"         "     45 (28.5) "          
#>   "edema (%)"               "    "                     
#>   "   0"                    "    132 (83.5) "          
#>   "   0.5"                  "     16 (10.1) "          
#>   "   1"                    "     10 ( 6.3) "          
#>   "bili (median [IQR])"     "   1.40 [0.80, 3.20]"     
#>   "chol (median [IQR])"     " 315.50 [247.75, 417.00]" 
#>   "albumin (mean (SD))"     "   3.52 (0.44)"           
#>   "copper (median [IQR])"   "  73.00 [40.00, 121.00]"  
#>   "alk.phos (median [IQR])" "1214.50 [840.75, 2028.00]"
#>   "ast (mean (SD))"         " 120.21 (54.52)"          
#>   "trig (median [IQR])"     " 106.00 [84.50, 146.00]"  
#>   "platelet (mean (SD))"    " 258.75 (100.32)"         
#>   "protime (mean (SD))"     "  10.65 (0.85)"           
#>   "stage (%)"               "    "                     
#>   "   1"                    "     12 ( 7.6) "          
#>   "   2"                    "     35 (22.2) "          
#>   "   3"                    "     56 (35.4) "          
#>   "   4"                    "     55 (34.8) "          
#>                            "Stratified by trt"
#>  ""                         "2"                         "p"      "test"   
#>   "n"                       "    154"                   ""       ""       
#>   "time (mean (SD))"        "1996.86 (1155.93)"         " 0.883" ""       
#>   "status (%)"              "    "                      " 0.884" "exact"  
#>   "   0"                    "     85 (55.2) "           ""       ""       
#>   "   1"                    "      9 ( 5.8) "           ""       ""       
#>   "   2"                    "     60 (39.0) "           ""       ""       
#>   "age (mean (SD))"         "  48.58 (9.96)"            " 0.018" ""       
#>   "sex = f (%)"             "    139 (90.3) "           " 0.421" ""       
#>   "ascites = 1 (%)"         "     10 ( 6.5) "           " 0.567" ""       
#>   "hepato = 1 (%)"          "     87 (56.5) "           " 0.088" ""       
#>   "spiders = 1 (%)"         "     45 (29.2) "           " 0.985" ""       
#>   "edema (%)"               "    "                      " 0.877" ""       
#>   "   0"                    "    131 (85.1) "           ""       ""       
#>   "   0.5"                  "     13 ( 8.4) "           ""       ""       
#>   "   1"                    "     10 ( 6.5) "           ""       ""       
#>   "bili (median [IQR])"     "   1.30 [0.72, 3.60]"      " 0.842" "nonnorm"
#>   "chol (median [IQR])"     " 303.50 [254.25, 377.00]"  " 0.544" "nonnorm"
#>   "albumin (mean (SD))"     "   3.52 (0.40)"            " 0.874" ""       
#>   "copper (median [IQR])"   "  73.00 [43.00, 139.00]"   " 0.717" "nonnorm"
#>   "alk.phos (median [IQR])" "1283.00 [922.50, 1949.75]" " 0.812" "nonnorm"
#>   "ast (mean (SD))"         " 124.97 (58.93)"           " 0.460" ""       
#>   "trig (median [IQR])"     " 113.00 [84.50, 155.00]"   " 0.370" "nonnorm"
#>   "platelet (mean (SD))"    " 265.20 (90.73)"           " 0.555" ""       
#>   "protime (mean (SD))"     "  10.80 (1.14)"            " 0.197" ""       
#>   "stage (%)"               "    "                      " 0.205" "exact"  
#>   "   1"                    "      4 ( 2.6) "           ""       ""       
#>   "   2"                    "     32 (20.8) "           ""       ""       
#>   "   3"                    "     64 (41.6) "           ""       ""       
#>   "   4"                    "     54 (35.1) "           ""       ""       

## If you want to center-align values in Word, use noSpaces option.
print(tableOne, nonnormal = c("bili","chol","copper","alk.phos","trig"),
      exact = c("status","stage"), quote = TRUE, noSpaces = TRUE)
#>                            "Stratified by trt"
#>  ""                         "1"                        
#>   "n"                       "158"                      
#>   "time (mean (SD))"        "2015.62 (1094.12)"        
#>   "status (%)"              ""                         
#>   "   0"                    "83 (52.5)"                
#>   "   1"                    "10 (6.3)"                 
#>   "   2"                    "65 (41.1)"                
#>   "age (mean (SD))"         "51.42 (11.01)"            
#>   "sex = f (%)"             "137 (86.7)"               
#>   "ascites = 1 (%)"         "14 (8.9)"                 
#>   "hepato = 1 (%)"          "73 (46.2)"                
#>   "spiders = 1 (%)"         "45 (28.5)"                
#>   "edema (%)"               ""                         
#>   "   0"                    "132 (83.5)"               
#>   "   0.5"                  "16 (10.1)"                
#>   "   1"                    "10 (6.3)"                 
#>   "bili (median [IQR])"     "1.40 [0.80, 3.20]"        
#>   "chol (median [IQR])"     "315.50 [247.75, 417.00]"  
#>   "albumin (mean (SD))"     "3.52 (0.44)"              
#>   "copper (median [IQR])"   "73.00 [40.00, 121.00]"    
#>   "alk.phos (median [IQR])" "1214.50 [840.75, 2028.00]"
#>   "ast (mean (SD))"         "120.21 (54.52)"           
#>   "trig (median [IQR])"     "106.00 [84.50, 146.00]"   
#>   "platelet (mean (SD))"    "258.75 (100.32)"          
#>   "protime (mean (SD))"     "10.65 (0.85)"             
#>   "stage (%)"               ""                         
#>   "   1"                    "12 (7.6)"                 
#>   "   2"                    "35 (22.2)"                
#>   "   3"                    "56 (35.4)"                
#>   "   4"                    "55 (34.8)"                
#>                            "Stratified by trt"
#>  ""                         "2"                         "p"     "test"   
#>   "n"                       "154"                       ""      ""       
#>   "time (mean (SD))"        "1996.86 (1155.93)"         "0.883" ""       
#>   "status (%)"              ""                          "0.884" "exact"  
#>   "   0"                    "85 (55.2)"                 ""      ""       
#>   "   1"                    "9 (5.8)"                   ""      ""       
#>   "   2"                    "60 (39.0)"                 ""      ""       
#>   "age (mean (SD))"         "48.58 (9.96)"              "0.018" ""       
#>   "sex = f (%)"             "139 (90.3)"                "0.421" ""       
#>   "ascites = 1 (%)"         "10 (6.5)"                  "0.567" ""       
#>   "hepato = 1 (%)"          "87 (56.5)"                 "0.088" ""       
#>   "spiders = 1 (%)"         "45 (29.2)"                 "0.985" ""       
#>   "edema (%)"               ""                          "0.877" ""       
#>   "   0"                    "131 (85.1)"                ""      ""       
#>   "   0.5"                  "13 (8.4)"                  ""      ""       
#>   "   1"                    "10 (6.5)"                  ""      ""       
#>   "bili (median [IQR])"     "1.30 [0.72, 3.60]"         "0.842" "nonnorm"
#>   "chol (median [IQR])"     "303.50 [254.25, 377.00]"   "0.544" "nonnorm"
#>   "albumin (mean (SD))"     "3.52 (0.40)"               "0.874" ""       
#>   "copper (median [IQR])"   "73.00 [43.00, 139.00]"     "0.717" "nonnorm"
#>   "alk.phos (median [IQR])" "1283.00 [922.50, 1949.75]" "0.812" "nonnorm"
#>   "ast (mean (SD))"         "124.97 (58.93)"            "0.460" ""       
#>   "trig (median [IQR])"     "113.00 [84.50, 155.00]"    "0.370" "nonnorm"
#>   "platelet (mean (SD))"    "265.20 (90.73)"            "0.555" ""       
#>   "protime (mean (SD))"     "10.80 (1.14)"              "0.197" ""       
#>   "stage (%)"               ""                          "0.205" "exact"  
#>   "   1"                    "4 (2.6)"                   ""      ""       
#>   "   2"                    "32 (20.8)"                 ""      ""       
#>   "   3"                    "64 (41.6)"                 ""      ""       
#>   "   4"                    "54 (35.1)"                 ""      ""       

## If SMDs are needed as numericals, use ExtractSmd()
ExtractSmd(tableOne)
#>                1 vs 2
#> time     0.0166658751
#> status   0.0537576290
#> age      0.2702619258
#> sex      0.1114116116
#> ascites  0.0890061759
#> hepato   0.2069941350
#> spiders  0.0163284440
#> edema    0.0581165880
#> bili     0.1710905651
#> chol     0.0382210537
#> albumin  0.0180021838
#> copper   0.0001200022
#> alk.phos 0.0365323630
#> ast      0.0837836058
#> trig     0.0170615337
#> platelet 0.0674763888
#> protime  0.1460939117
#> stage    0.2460083410