Format regression results in medically decent format

It shows the regression result in the HR [95% CI] p-value format, which is usually the form used in medical research papers.

ShowRegTable(
  model,
  exp = TRUE,
  digits = 2,
  pDigits = 3,
  printToggle = TRUE,
  quote = FALSE,
  ciFun = confint
)

Arguments

model

Regression model result objects that have the summary and confint methods.

exp

TRUE by default. You need to specify exp = FALSE if your model is has the indentity link function (linear regression, etc).

digits

Number of digits to print for the main part.

pDigits

Number of digits to print for the p-values.

printToggle

Whether to print the output. If FALSE, no output is created, and a matrix is invisibly returned.

quote

Whether to show everything in quotes. The default is FALSE. If TRUE, everything including the row and column names are quoted so that you can copy it to Excel easily.

ciFun

Function used for calculation. confint is the default. For generalized linear models this gives the profile likelihood-based calculation, which may take too much time for large models, use confint.default for simple normal approximation method (+/- 1.96 * standard error).

Value

A matrix containing what you see is returned invisibly. You can capture it by assignment to an object.

Author

Kazuki Yoshida

Examples

## Load
library(tableone)

## Load Mayo Clinic Primary Biliary Cirrhosis Data
library(survival)
data(pbc)
## Check variables
head(pbc)
#>   id time status trt      age sex ascites hepato spiders edema bili chol
#> 1  1  400      2   1 58.76523   f       1      1       1   1.0 14.5  261
#> 2  2 4500      0   1 56.44627   f       0      1       1   0.0  1.1  302
#> 3  3 1012      2   1 70.07255   m       0      0       0   0.5  1.4  176
#> 4  4 1925      2   1 54.74059   f       0      1       1   0.5  1.8  244
#> 5  5 1504      1   2 38.10541   f       0      1       1   0.0  3.4  279
#> 6  6 2503      2   2 66.25873   f       0      1       0   0.0  0.8  248
#>   albumin copper alk.phos    ast trig platelet protime stage
#> 1    2.60    156   1718.0 137.95  172      190    12.2     4
#> 2    4.14     54   7394.8 113.52   88      221    10.6     3
#> 3    3.48    210    516.0  96.10   55      151    12.0     4
#> 4    2.54     64   6121.8  60.63   92      183    10.3     4
#> 5    3.53    143    671.0 113.15   72      136    10.9     3
#> 6    3.98     50    944.0  93.00   63       NA    11.0     3

## Fit a Cox regression model
objCoxph <- coxph(formula = Surv(time, status == 2) ~ trt + age + albumin + ascites,
                  data    = pbc)

## Show the simple table
ShowRegTable(objCoxph)
#>         exp(coef) [confint] p     
#> trt     1.03 [0.72, 1.47]    0.891
#> age     1.03 [1.01, 1.05]    0.005
#> albumin 0.26 [0.17, 0.41]   <0.001
#> ascites 2.93 [1.68, 5.11]   <0.001

## Show with quote to ease copy and paste
ShowRegTable(objCoxph, quote = TRUE)
#>           "exp(coef) [confint]" "p"     
#> "trt"     "1.03 [0.72, 1.47]"   " 0.891"
#> "age"     "1.03 [1.01, 1.05]"   " 0.005"
#> "albumin" "0.26 [0.17, 0.41]"   "<0.001"
#> "ascites" "2.93 [1.68, 5.11]"   "<0.001"