Format and print ContTable class objects

print method for the ContTable class objects created by CreateContTable function.

# S3 method for ContTable
print(
  x,
  digits = 2,
  pDigits = 3,
  quote = FALSE,
  missing = FALSE,
  explain = TRUE,
  printToggle = TRUE,
  noSpaces = FALSE,
  nonnormal = NULL,
  minMax = FALSE,
  insertLevel = FALSE,
  test = TRUE,
  smd = FALSE,
  formatOptions = list(scientific = FALSE),
  ...
)

Arguments

x

Object returned by CreateContTable function.

digits

Number of digits to print in the table.

pDigits

Number of digits to print for p-values (also used for standardized mean differences).

quote

Whether to show everything in quotes. The default is FALSE. If TRUE, everything including the row and column names are quoted so that you can copy it to Excel easily.

missing

Whether to show missing data information.

explain

Whether to add explanation to the variable names, i.e., (mean (SD) or median [IQR]) is added to the variable names.

printToggle

Whether to print the output. If FALSE, no output is created, and a matrix is invisibly returned.

noSpaces

Whether to remove spaces added for alignment. Use this option if you prefer to align numbers yourself in other software.

nonnormal

A character vector to specify the variables for which the p-values should be those of nonparametric tests. By default all p-values are from normal assumption-based tests (oneway.test).

minMax

Whether to use [min,max] instead of [p25,p75] for nonnormal variables. The default is FALSE.

insertLevel

Whether to add an empty level column to the left of strata.

test

Whether to show p-values. TRUE by default. If FALSE, only the numerical summaries are shown.

smd

Whether to show standardized mean differences. FALSE by default. If there are more than one contrasts, the average of all possible standardized mean differences is shown. For individual contrasts, use summary.

formatOptions

A list of options, which will be passed to format. Can be used to modify the big.mark, decimal.mark, big.interval etc. The default is list(scientific = FALSE). The options digits, nsmall, justify and trim are not available. (Experimental)

...

For compatibility with generic. Ignored.

Value

A matrix object containing what you see is also invisibly returned. This can be assinged a name and exported via write.csv.

See also

Author

Kazuki Yoshida, Alexander Bartel

Examples

## Load
library(tableone)

## Load Mayo Clinic Primary Biliary Cirrhosis Data
library(survival)
data(pbc)
## Check variables
head(pbc)
#>   id time status trt      age sex ascites hepato spiders edema bili chol
#> 1  1  400      2   1 58.76523   f       1      1       1   1.0 14.5  261
#> 2  2 4500      0   1 56.44627   f       0      1       1   0.0  1.1  302
#> 3  3 1012      2   1 70.07255   m       0      0       0   0.5  1.4  176
#> 4  4 1925      2   1 54.74059   f       0      1       1   0.5  1.8  244
#> 5  5 1504      1   2 38.10541   f       0      1       1   0.0  3.4  279
#> 6  6 2503      2   2 66.25873   f       0      1       0   0.0  0.8  248
#>   albumin copper alk.phos    ast trig platelet protime stage
#> 1    2.60    156   1718.0 137.95  172      190    12.2     4
#> 2    4.14     54   7394.8 113.52   88      221    10.6     3
#> 3    3.48    210    516.0  96.10   55      151    12.0     4
#> 4    2.54     64   6121.8  60.63   92      183    10.3     4
#> 5    3.53    143    671.0 113.15   72      136    10.9     3
#> 6    3.98     50    944.0  93.00   63       NA    11.0     3

## Create an overall table for continuous variables
contVars <- c("time","age","bili","chol","albumin","copper",
              "alk.phos","ast","trig","platelet","protime")
contTableOverall <- CreateContTable(vars = contVars, data = pbc)

## Simply typing the object name will invoke the print.ContTable method,
## which will show the sample size, means and standard deviations.
contTableOverall
#>                       
#>                        Overall          
#>   n                    418              
#>   time (mean (SD))     1917.78 (1104.67)
#>   age (mean (SD))        50.74 (10.45)  
#>   bili (mean (SD))        3.22 (4.41)   
#>   chol (mean (SD))      369.51 (231.94) 
#>   albumin (mean (SD))     3.50 (0.42)   
#>   copper (mean (SD))     97.65 (85.61)  
#>   alk.phos (mean (SD)) 1982.66 (2140.39)
#>   ast (mean (SD))       122.56 (56.70)  
#>   trig (mean (SD))      124.70 (65.15)  
#>   platelet (mean (SD))  257.02 (98.33)  
#>   protime (mean (SD))    10.73 (1.02)   

## To further examine the variables, use the summary.ContTable method,
## which will show more details.
summary(contTableOverall)
#> strata: Overall
#>            n miss p.miss   mean      sd median    p25    p75   min     max
#> time     418    0   0.00 1917.8 1104.67 1730.0 1092.8 2613.5  41.0  4795.0
#> age      418    0   0.00   50.7   10.45   51.0   42.8   58.2  26.3    78.4
#> bili     418    0   0.00    3.2    4.41    1.4    0.8    3.4   0.3    28.0
#> chol     418  134  32.06  369.5  231.94  309.5  249.5  400.0 120.0  1775.0
#> albumin  418    0   0.00    3.5    0.42    3.5    3.2    3.8   2.0     4.6
#> copper   418  108  25.84   97.6   85.61   73.0   41.2  123.0   4.0   588.0
#> alk.phos 418  106  25.36 1982.7 2140.39 1259.0  871.5 1980.0 289.0 13862.4
#> ast      418  106  25.36  122.6   56.70  114.7   80.6  151.9  26.4   457.2
#> trig     418  136  32.54  124.7   65.15  108.0   84.2  151.0  33.0   598.0
#> platelet 418   11   2.63  257.0   98.33  251.0  188.5  318.0  62.0   721.0
#> protime  418    2   0.48   10.7    1.02   10.6   10.0   11.1   9.0    18.0
#>            skew  kurt
#> time      0.473 -0.48
#> age       0.087 -0.62
#> bili      2.718  8.07
#> chol      3.409 14.34
#> albumin  -0.468  0.57
#> copper    2.304  7.62
#> alk.phos  2.993  9.66
#> ast       1.449  4.31
#> trig      2.524 11.80
#> platelet  0.627  0.86
#> protime   2.223 10.04

## c("age","chol","copper","alk.phos","trig","protime") appear highly skewed.
## Specify them in the nonnormal argument, and the display changes to the median,
## and the [25th, 75th] percentile.
nonNormalVars <- c("age","chol","copper","alk.phos","trig","protime")
print(contTableOverall, nonnormal = nonNormalVars)
#>                          
#>                           Overall                  
#>   n                       418                      
#>   time (mean (SD))        1917.78 (1104.67)        
#>   age (median [IQR])        51.00 [42.83, 58.24]   
#>   bili (mean (SD))           3.22 (4.41)           
#>   chol (median [IQR])      309.50 [249.50, 400.00] 
#>   albumin (mean (SD))        3.50 (0.42)           
#>   copper (median [IQR])     73.00 [41.25, 123.00]  
#>   alk.phos (median [IQR]) 1259.00 [871.50, 1980.00]
#>   ast (mean (SD))          122.56 (56.70)          
#>   trig (median [IQR])      108.00 [84.25, 151.00]  
#>   platelet (mean (SD))     257.02 (98.33)          
#>   protime (median [IQR])    10.60 [10.00, 11.10]   

## To show median [min,max] for nonnormal variables, use minMax = TRUE
print(contTableOverall, nonnormal = nonNormalVars, minMax = TRUE)
#>                            
#>                             Overall                   
#>   n                         418                       
#>   time (mean (SD))          1917.78 (1104.67)         
#>   age (median [range])        51.00 [26.28, 78.44]    
#>   bili (mean (SD))             3.22 (4.41)            
#>   chol (median [range])      309.50 [120.00, 1775.00] 
#>   albumin (mean (SD))          3.50 (0.42)            
#>   copper (median [range])     73.00 [4.00, 588.00]    
#>   alk.phos (median [range]) 1259.00 [289.00, 13862.40]
#>   ast (mean (SD))            122.56 (56.70)           
#>   trig (median [range])      108.00 [33.00, 598.00]   
#>   platelet (mean (SD))       257.02 (98.33)           
#>   protime (median [range])    10.60 [9.00, 18.00]     

## The table can be stratified by one or more variables
contTableBySexTrt <- CreateContTable(vars = contVars,
                                     strata = c("sex","trt"), data = pbc)

## print now includes p-values which are by default calculated by oneway.test (t-test
## equivalent in the two group case). It is formatted at the decimal place specified
## by the pDigits argument (3 by default). It does <0.001 for you.
contTableBySexTrt
#>                       Stratified by sex:trt
#>                        m:1               f:1               m:2              
#>   n                    21                137               15               
#>   time (mean (SD))     1793.48 (1244.70) 2049.67 (1070.20) 2156.33 (1428.56)
#>   age (mean (SD))        55.57 (12.61)     50.78 (10.65)     57.09 (10.07)  
#>   bili (mean (SD))        2.98 (2.11)       2.86 (3.81)       2.72 (2.46)   
#>   chol (mean (SD))      403.43 (204.95)   358.24 (210.46)   301.00 (111.32) 
#>   albumin (mean (SD))     3.63 (0.43)       3.50 (0.44)       3.50 (0.49)   
#>   copper (mean (SD))    174.90 (105.33)    85.71 (82.27)    125.40 (89.18)  
#>   alk.phos (mean (SD)) 2486.16 (2385.32) 1950.04 (2151.35) 1734.45 (2478.07)
#>   ast (mean (SD))       128.71 (48.74)    118.91 (55.40)    112.58 (44.38)  
#>   trig (mean (SD))      145.81 (56.62)    120.28 (73.41)    114.86 (39.59)  
#>   platelet (mean (SD))  232.65 (97.38)    262.59 (100.53)   240.13 (73.93)  
#>   protime (mean (SD))    10.84 (0.94)      10.62 (0.84)      11.23 (0.97)   
#>                       Stratified by sex:trt
#>                        f:2               p      test
#>   n                    139                          
#>   time (mean (SD))     1979.65 (1127.52)  0.730     
#>   age (mean (SD))        47.66 (9.54)    <0.001     
#>   bili (mean (SD))        3.75 (5.50)     0.394     
#>   chol (mean (SD))      381.73 (262.26)   0.512     
#>   albumin (mean (SD))     3.53 (0.39)     0.585     
#>   copper (mean (SD))     94.64 (79.25)   <0.001     
#>   alk.phos (mean (SD)) 1965.52 (2066.15)  0.706     
#>   ast (mean (SD))       126.30 (60.27)    0.598     
#>   trig (mean (SD))      126.38 (60.22)    0.370     
#>   platelet (mean (SD))  267.95 (92.20)    0.362     
#>   protime (mean (SD))    10.75 (1.15)     0.137     

## The nonnormal argument toggles the p-values to the nonparametric result from
## kruskal.test (wilcox.test equivalent for the two group case).
print(contTableBySexTrt, nonnormal = nonNormalVars)
#>                          Stratified by sex:trt
#>                           m:1                       f:1                      
#>   n                       21                        137                      
#>   time (mean (SD))        1793.48 (1244.70)         2049.67 (1070.20)        
#>   age (median [IQR])        55.99 [46.45, 65.76]      51.23 [42.51, 57.04]   
#>   bili (mean (SD))           2.98 (2.11)               2.86 (3.81)           
#>   chol (median [IQR])      376.00 [251.00, 456.00]   309.00 [247.50, 399.50] 
#>   albumin (mean (SD))        3.63 (0.43)               3.50 (0.44)           
#>   copper (median [IQR])    158.00 [100.00, 225.00]    67.00 [36.75, 105.00]  
#>   alk.phos (median [IQR]) 1664.00 [983.00, 2496.00] 1212.00 [823.00, 1877.00]
#>   ast (mean (SD))          128.71 (48.74)            118.91 (55.40)          
#>   trig (median [IQR])      142.00 [107.00, 194.00]   101.00 [84.00, 134.00]  
#>   platelet (mean (SD))     232.65 (97.38)            262.59 (100.53)         
#>   protime (median [IQR])    10.80 [10.30, 11.00]      10.60 [10.00, 11.00]   
#>                          Stratified by sex:trt
#>                           m:2                       f:2                      
#>   n                       15                        139                      
#>   time (mean (SD))        2156.33 (1428.56)         1979.65 (1127.52)        
#>   age (median [IQR])        52.83 [49.45, 65.80]      47.18 [41.26, 55.46]   
#>   bili (mean (SD))           2.72 (2.46)               3.75 (5.50)           
#>   chol (median [IQR])      259.00 [235.75, 342.25]   306.50 [258.00, 381.00] 
#>   albumin (mean (SD))        3.50 (0.49)               3.53 (0.39)           
#>   copper (median [IQR])     84.00 [52.50, 198.00]     69.50 [42.50, 127.25]  
#>   alk.phos (median [IQR]) 1070.00 [737.00, 1211.50] 1345.00 [941.00, 1992.00]
#>   ast (mean (SD))          112.58 (44.38)            126.30 (60.27)          
#>   trig (median [IQR])      108.50 [90.25, 137.00]    113.00 [84.00, 157.00]  
#>   platelet (mean (SD))     240.13 (73.93)            267.95 (92.20)          
#>   protime (median [IQR])    11.20 [10.55, 11.55]      10.60 [9.90, 11.25]    
#>                          Stratified by sex:trt
#>                           p      test   
#>   n                                     
#>   time (mean (SD))         0.730        
#>   age (median [IQR])       0.001 nonnorm
#>   bili (mean (SD))         0.394        
#>   chol (median [IQR])      0.458 nonnorm
#>   albumin (mean (SD))      0.585        
#>   copper (median [IQR])   <0.001 nonnorm
#>   alk.phos (median [IQR])  0.180 nonnorm
#>   ast (mean (SD))          0.598        
#>   trig (median [IQR])      0.058 nonnorm
#>   platelet (mean (SD))     0.362        
#>   protime (median [IQR])   0.081 nonnorm

## The minMax argument toggles whether to show median [range]
print(contTableBySexTrt, nonnormal = nonNormalVars, minMax = TRUE)
#>                            Stratified by sex:trt
#>                             m:1                       
#>   n                         21                        
#>   time (mean (SD))          1793.48 (1244.70)         
#>   age (median [range])        55.99 [33.48, 78.44]    
#>   bili (mean (SD))             2.98 (2.11)            
#>   chol (median [range])      376.00 [168.00, 1000.00] 
#>   albumin (mean (SD))          3.63 (0.43)            
#>   copper (median [range])    158.00 [25.00, 444.00]   
#>   alk.phos (median [range]) 1664.00 [516.00, 10165.00]
#>   ast (mean (SD))            128.71 (48.74)           
#>   trig (median [range])      142.00 [55.00, 242.00]   
#>   platelet (mean (SD))       232.65 (97.38)           
#>   protime (median [range])    10.80 [9.70, 14.10]     
#>                            Stratified by sex:trt
#>                             f:1                       
#>   n                         137                       
#>   time (mean (SD))          2049.67 (1070.20)         
#>   age (median [range])        51.23 [26.28, 76.71]    
#>   bili (mean (SD))             2.86 (3.81)            
#>   chol (median [range])      309.00 [127.00, 1712.00] 
#>   albumin (mean (SD))          3.50 (0.44)            
#>   copper (median [range])     67.00 [9.00, 588.00]    
#>   alk.phos (median [range]) 1212.00 [369.00, 11552.00]
#>   ast (mean (SD))            118.91 (55.40)           
#>   trig (median [range])      101.00 [33.00, 598.00]   
#>   platelet (mean (SD))       262.59 (100.53)          
#>   protime (median [range])    10.60 [9.00, 13.00]     
#>                            Stratified by sex:trt
#>                             m:2                       
#>   n                         15                        
#>   time (mean (SD))          2156.33 (1428.56)         
#>   age (median [range])        52.83 [43.90, 74.52]    
#>   bili (mean (SD))             2.72 (2.46)            
#>   chol (median [range])      259.00 [151.00, 546.00]  
#>   albumin (mean (SD))          3.50 (0.49)            
#>   copper (median [range])     84.00 [13.00, 281.00]   
#>   alk.phos (median [range]) 1070.00 [577.00, 10396.80]
#>   ast (mean (SD))            112.58 (44.38)           
#>   trig (median [range])      108.50 [49.00, 188.00]   
#>   platelet (mean (SD))       240.13 (73.93)           
#>   protime (median [range])    11.20 [10.00, 13.20]    
#>                            Stratified by sex:trt
#>                             f:2                        p      test   
#>   n                         139                                      
#>   time (mean (SD))          1979.65 (1127.52)           0.730        
#>   age (median [range])        47.18 [30.57, 70.56]      0.001 nonnorm
#>   bili (mean (SD))             3.75 (5.50)              0.394        
#>   chol (median [range])      306.50 [120.00, 1775.00]   0.458 nonnorm
#>   albumin (mean (SD))          3.53 (0.39)              0.585        
#>   copper (median [range])     69.50 [4.00, 558.00]     <0.001 nonnorm
#>   alk.phos (median [range]) 1345.00 [289.00, 13862.40]  0.180 nonnorm
#>   ast (mean (SD))            126.30 (60.27)             0.598        
#>   trig (median [range])      113.00 [44.00, 432.00]     0.058 nonnorm
#>   platelet (mean (SD))       267.95 (92.20)             0.362        
#>   protime (median [range])    10.60 [9.20, 17.10]       0.081 nonnorm

## summary now includes both types of p-values
summary(contTableBySexTrt)
#> sex: m
#> trt: 1
#>           n miss p.miss   mean      sd median   p25    p75   min     max   skew
#> time     21    0    0.0 1793.5 1244.70 1302.0 999.0 2386.0 140.0  4459.0  1.042
#> age      21    0    0.0   55.6   12.61   56.0  46.5   65.8  33.5    78.4 -0.200
#> bili     21    0    0.0    3.0    2.11    2.3   1.4    4.0   0.6     7.3  1.019
#> chol     21    0    0.0  403.4  204.95  376.0 251.0  456.0 168.0  1000.0  1.329
#> albumin  21    0    0.0    3.6    0.43    3.8   3.4    3.9   2.6     4.2 -0.966
#> copper   21    0    0.0  174.9  105.33  158.0 100.0  225.0  25.0   444.0  1.082
#> alk.phos 21    0    0.0 2486.2 2385.32 1664.0 983.0 2496.0 516.0 10165.0  2.051
#> ast      21    0    0.0  128.7   48.74  127.0  86.0  158.1  56.8   221.9  0.309
#> trig     21    0    0.0  145.8   56.62  142.0 107.0  194.0  55.0   242.0  0.164
#> platelet 21    1    4.8  232.7   97.38  227.5 148.2  313.8  70.0   394.0 -0.045
#> protime  21    0    0.0   10.8    0.94   10.8  10.3   11.0   9.7    14.1  2.113
#>           kurt
#> time      0.10
#> age      -0.71
#> bili     -0.12
#> chol      2.15
#> albumin   0.34
#> copper    0.97
#> alk.phos  4.49
#> ast      -0.86
#> trig     -1.04
#> platelet -1.16
#> protime   6.84
#> ------------------------------------------------------------ 
#> sex: f
#> trt: 1
#>            n miss p.miss   mean      sd median    p25    p75   min     max
#> time     137    0   0.00 2049.7 1070.20 1945.0 1293.0 2644.0  41.0  4556.0
#> age      137    0   0.00   50.8   10.65   51.2   42.5   57.0  26.3    76.7
#> bili     137    0   0.00    2.9    3.81    1.3    0.8    3.2   0.3    20.0
#> chol     137   18  13.14  358.2  210.46  309.0  247.5  399.5 127.0  1712.0
#> albumin  137    0   0.00    3.5    0.44    3.5    3.2    3.8   2.1     4.6
#> copper   137    1   0.73   85.7   82.27   67.0   36.8  105.0   9.0   588.0
#> alk.phos 137    0   0.00 1950.0 2151.35 1212.0  823.0 1877.0 369.0 11552.0
#> ast      137    0   0.00  118.9   55.40  108.5   74.4  150.0  26.4   338.0
#> trig     137   19  13.87  120.3   73.41  101.0   84.0  134.0  33.0   598.0
#> platelet 137    1   0.73  262.6  100.53  258.5  194.2  322.0  62.0   563.0
#> protime  137    0   0.00   10.6    0.84   10.6   10.0   11.0   9.0    13.0
#>            skew  kurt
#> time      0.318 -0.41
#> age       0.045 -0.48
#> bili      2.652  7.10
#> chol      4.281 23.79
#> albumin  -0.327  0.42
#> copper    3.181 13.92
#> alk.phos  2.876  8.39
#> ast       1.196  1.91
#> trig      3.278 16.03
#> platelet  0.565  0.34
#> protime   0.916  0.57
#> ------------------------------------------------------------ 
#> sex: m
#> trt: 2
#>           n miss p.miss   mean      sd median    p25    p75   min     max  skew
#> time     15    0    0.0 2156.3 1428.56 1656.0 1053.5 3420.0 191.0  4427.0  0.30
#> age      15    0    0.0   57.1   10.07   52.8   49.4   65.8  43.9    74.5  0.45
#> bili     15    0    0.0    2.7    2.46    1.9    1.1    2.9   0.6     8.6  1.54
#> chol     15    1    6.7  301.0  111.32  259.0  235.8  342.2 151.0   546.0  0.91
#> albumin  15    0    0.0    3.5    0.49    3.6    3.3    3.8   2.3     4.3 -0.96
#> copper   15    0    0.0  125.4   89.18   84.0   52.5  198.0  13.0   281.0  0.50
#> alk.phos 15    0    0.0 1734.5 2478.07 1070.0  737.0 1211.5 577.0 10396.8  3.49
#> ast      15    0    0.0  112.6   44.38  110.0   74.2  129.4  46.5   188.3  0.32
#> trig     15    1    6.7  114.9   39.59  108.5   90.2  137.0  49.0   188.0  0.44
#> platelet 15    0    0.0  240.1   73.93  214.0  189.5  294.5 119.0   360.0  0.18
#> protime  15    0    0.0   11.2    0.97   11.2   10.6   11.6  10.0    13.2  0.72
#>           kurt
#> time     -1.39
#> age      -1.29
#> bili      1.43
#> chol      0.36
#> albumin   1.88
#> copper   -1.13
#> alk.phos 12.68
#> ast      -0.80
#> trig     -0.36
#> platelet -1.02
#> protime   0.18
#> ------------------------------------------------------------ 
#> sex: f
#> trt: 2
#>            n miss p.miss   mean      sd median    p25    p75   min     max
#> time     139    0   0.00 1979.7 1127.52 1832.0 1157.0 2717.0  51.0  4523.0
#> age      139    0   0.00   47.7    9.54   47.2   41.3   55.5  30.6    70.6
#> bili     139    0   0.00    3.7    5.50    1.3    0.7    3.6   0.3    28.0
#> chol     139    9   6.47  381.7  262.26  306.5  258.0  381.0 120.0  1775.0
#> albumin  139    0   0.00    3.5    0.39    3.5    3.3    3.8   2.0     4.4
#> copper   139    1   0.72   94.6   79.25   69.5   42.5  127.2   4.0   558.0
#> alk.phos 139    0   0.00 1965.5 2066.15 1345.0  941.0 1992.0 289.0 13862.4
#> ast      139    0   0.00  126.3   60.27  117.8   84.5  152.7  28.4   457.2
#> trig     139   10   7.19  126.4   60.22  113.0   84.0  157.0  44.0   432.0
#> platelet 139    2   1.44  267.9   92.20  265.0  213.0  324.0  71.0   487.0
#> protime  139    0   0.00   10.8    1.15   10.6    9.9   11.2   9.2    17.1
#>           skew  kurt
#> time      0.34 -0.64
#> age       0.16 -0.74
#> bili      2.57  6.54
#> chol      3.01 10.13
#> albumin  -0.83  2.06
#> copper    2.25  8.30
#> alk.phos  3.39 13.57
#> ast       1.76  6.24
#> trig      1.71  5.27
#> platelet  0.21 -0.36
#> protime   2.02  7.15
#> 
#> p-values
#>               pNormal   pNonNormal
#> time     7.300496e-01 5.566440e-01
#> age      1.286381e-04 6.517437e-04
#> bili     3.937430e-01 2.049525e-01
#> chol     5.122223e-01 4.578773e-01
#> albumin  5.853926e-01 4.041206e-01
#> copper   6.237989e-05 4.712924e-05
#> alk.phos 7.063812e-01 1.801480e-01
#> ast      5.982898e-01 5.086616e-01
#> trig     3.701073e-01 5.827726e-02
#> platelet 3.624740e-01 4.451277e-01
#> protime  1.365463e-01 8.114383e-02
#> 
#> Standardize mean differences
#>            average     1 vs 2     1 vs 3     1 vs 4     2 vs 3     2 vs 4
#> time     0.1556365 0.22071813 0.27083134 0.15677410 0.08450703 0.06369595
#> age      0.5214974 0.40993004 0.13353519 0.70700091 0.60864596 0.30853876
#> bili     0.1352025 0.04023105 0.11373946 0.18444097 0.04265701 0.18855913
#> chol     0.2950579 0.21756422 0.62108339 0.09219212 0.33996564 0.09881382
#> albumin  0.1637683 0.30878937 0.28262444 0.26228013 0.01074465 0.06658036
#> copper   0.5416574 0.94378898 0.50727571 0.86116098 0.46259003 0.11048709
#> alk.phos 0.1633259 0.23603622 0.30907333 0.23332162 0.09290687 0.00733728
#> ast      0.1818681 0.18797132 0.34614452 0.04399580 0.12602823 0.12777729
#> trig     0.2940632 0.38944815 0.63360898 0.33242327 0.09194315 0.09085446
#> platelet 0.2340433 0.30251429 0.08655951 0.37226635 0.25448008 0.05557853
#> protime  0.3283721 0.24596092 0.40162254 0.08476002 0.66430044 0.12924857
#>              3 vs 4
#> time     0.13729269
#> age      0.96133382
#> bili     0.24158721
#> chol     0.40072828
#> albumin  0.05159074
#> copper   0.36464167
#> alk.phos 0.10128012
#> ast      0.25929154
#> trig     0.22610097
#> platelet 0.33286083
#> protime  0.44433997

## If your work flow includes copying to Excel and Word when writing manuscripts,
## you may benefit from the quote argument. This will quote everything so that
## Excel does not mess up the cells.
print(contTableBySexTrt, nonnormal = nonNormalVars, quote = TRUE)
#>                            "Stratified by sex:trt"
#>  ""                         "m:1"                      
#>   "n"                       "21"                       
#>   "time (mean (SD))"        "1793.48 (1244.70)"        
#>   "age (median [IQR])"      "  55.99 [46.45, 65.76]"   
#>   "bili (mean (SD))"        "   2.98 (2.11)"           
#>   "chol (median [IQR])"     " 376.00 [251.00, 456.00]" 
#>   "albumin (mean (SD))"     "   3.63 (0.43)"           
#>   "copper (median [IQR])"   " 158.00 [100.00, 225.00]" 
#>   "alk.phos (median [IQR])" "1664.00 [983.00, 2496.00]"
#>   "ast (mean (SD))"         " 128.71 (48.74)"          
#>   "trig (median [IQR])"     " 142.00 [107.00, 194.00]" 
#>   "platelet (mean (SD))"    " 232.65 (97.38)"          
#>   "protime (median [IQR])"  "  10.80 [10.30, 11.00]"   
#>                            "Stratified by sex:trt"
#>  ""                         "f:1"                      
#>   "n"                       "137"                      
#>   "time (mean (SD))"        "2049.67 (1070.20)"        
#>   "age (median [IQR])"      "  51.23 [42.51, 57.04]"   
#>   "bili (mean (SD))"        "   2.86 (3.81)"           
#>   "chol (median [IQR])"     " 309.00 [247.50, 399.50]" 
#>   "albumin (mean (SD))"     "   3.50 (0.44)"           
#>   "copper (median [IQR])"   "  67.00 [36.75, 105.00]"  
#>   "alk.phos (median [IQR])" "1212.00 [823.00, 1877.00]"
#>   "ast (mean (SD))"         " 118.91 (55.40)"          
#>   "trig (median [IQR])"     " 101.00 [84.00, 134.00]"  
#>   "platelet (mean (SD))"    " 262.59 (100.53)"         
#>   "protime (median [IQR])"  "  10.60 [10.00, 11.00]"   
#>                            "Stratified by sex:trt"
#>  ""                         "m:2"                      
#>   "n"                       "15"                       
#>   "time (mean (SD))"        "2156.33 (1428.56)"        
#>   "age (median [IQR])"      "  52.83 [49.45, 65.80]"   
#>   "bili (mean (SD))"        "   2.72 (2.46)"           
#>   "chol (median [IQR])"     " 259.00 [235.75, 342.25]" 
#>   "albumin (mean (SD))"     "   3.50 (0.49)"           
#>   "copper (median [IQR])"   "  84.00 [52.50, 198.00]"  
#>   "alk.phos (median [IQR])" "1070.00 [737.00, 1211.50]"
#>   "ast (mean (SD))"         " 112.58 (44.38)"          
#>   "trig (median [IQR])"     " 108.50 [90.25, 137.00]"  
#>   "platelet (mean (SD))"    " 240.13 (73.93)"          
#>   "protime (median [IQR])"  "  11.20 [10.55, 11.55]"   
#>                            "Stratified by sex:trt"
#>  ""                         "f:2"                       "p"      "test"   
#>   "n"                       "139"                       ""       ""       
#>   "time (mean (SD))"        "1979.65 (1127.52)"         " 0.730" ""       
#>   "age (median [IQR])"      "  47.18 [41.26, 55.46]"    " 0.001" "nonnorm"
#>   "bili (mean (SD))"        "   3.75 (5.50)"            " 0.394" ""       
#>   "chol (median [IQR])"     " 306.50 [258.00, 381.00]"  " 0.458" "nonnorm"
#>   "albumin (mean (SD))"     "   3.53 (0.39)"            " 0.585" ""       
#>   "copper (median [IQR])"   "  69.50 [42.50, 127.25]"   "<0.001" "nonnorm"
#>   "alk.phos (median [IQR])" "1345.00 [941.00, 1992.00]" " 0.180" "nonnorm"
#>   "ast (mean (SD))"         " 126.30 (60.27)"           " 0.598" ""       
#>   "trig (median [IQR])"     " 113.00 [84.00, 157.00]"   " 0.058" "nonnorm"
#>   "platelet (mean (SD))"    " 267.95 (92.20)"           " 0.362" ""       
#>   "protime (median [IQR])"  "  10.60 [9.90, 11.25]"     " 0.081" "nonnorm"

## If you want to center-align values in Word, use noSpaces option.
print(contTableBySexTrt, nonnormal = nonNormalVars, quote = TRUE, noSpaces = TRUE)
#>                            "Stratified by sex:trt"
#>  ""                         "m:1"                      
#>   "n"                       "21"                       
#>   "time (mean (SD))"        "1793.48 (1244.70)"        
#>   "age (median [IQR])"      "55.99 [46.45, 65.76]"     
#>   "bili (mean (SD))"        "2.98 (2.11)"              
#>   "chol (median [IQR])"     "376.00 [251.00, 456.00]"  
#>   "albumin (mean (SD))"     "3.63 (0.43)"              
#>   "copper (median [IQR])"   "158.00 [100.00, 225.00]"  
#>   "alk.phos (median [IQR])" "1664.00 [983.00, 2496.00]"
#>   "ast (mean (SD))"         "128.71 (48.74)"           
#>   "trig (median [IQR])"     "142.00 [107.00, 194.00]"  
#>   "platelet (mean (SD))"    "232.65 (97.38)"           
#>   "protime (median [IQR])"  "10.80 [10.30, 11.00]"     
#>                            "Stratified by sex:trt"
#>  ""                         "f:1"                      
#>   "n"                       "137"                      
#>   "time (mean (SD))"        "2049.67 (1070.20)"        
#>   "age (median [IQR])"      "51.23 [42.51, 57.04]"     
#>   "bili (mean (SD))"        "2.86 (3.81)"              
#>   "chol (median [IQR])"     "309.00 [247.50, 399.50]"  
#>   "albumin (mean (SD))"     "3.50 (0.44)"              
#>   "copper (median [IQR])"   "67.00 [36.75, 105.00]"    
#>   "alk.phos (median [IQR])" "1212.00 [823.00, 1877.00]"
#>   "ast (mean (SD))"         "118.91 (55.40)"           
#>   "trig (median [IQR])"     "101.00 [84.00, 134.00]"   
#>   "platelet (mean (SD))"    "262.59 (100.53)"          
#>   "protime (median [IQR])"  "10.60 [10.00, 11.00]"     
#>                            "Stratified by sex:trt"
#>  ""                         "m:2"                      
#>   "n"                       "15"                       
#>   "time (mean (SD))"        "2156.33 (1428.56)"        
#>   "age (median [IQR])"      "52.83 [49.45, 65.80]"     
#>   "bili (mean (SD))"        "2.72 (2.46)"              
#>   "chol (median [IQR])"     "259.00 [235.75, 342.25]"  
#>   "albumin (mean (SD))"     "3.50 (0.49)"              
#>   "copper (median [IQR])"   "84.00 [52.50, 198.00]"    
#>   "alk.phos (median [IQR])" "1070.00 [737.00, 1211.50]"
#>   "ast (mean (SD))"         "112.58 (44.38)"           
#>   "trig (median [IQR])"     "108.50 [90.25, 137.00]"   
#>   "platelet (mean (SD))"    "240.13 (73.93)"           
#>   "protime (median [IQR])"  "11.20 [10.55, 11.55]"     
#>                            "Stratified by sex:trt"
#>  ""                         "f:2"                       "p"      "test"   
#>   "n"                       "139"                       ""       ""       
#>   "time (mean (SD))"        "1979.65 (1127.52)"         "0.730"  ""       
#>   "age (median [IQR])"      "47.18 [41.26, 55.46]"      "0.001"  "nonnorm"
#>   "bili (mean (SD))"        "3.75 (5.50)"               "0.394"  ""       
#>   "chol (median [IQR])"     "306.50 [258.00, 381.00]"   "0.458"  "nonnorm"
#>   "albumin (mean (SD))"     "3.53 (0.39)"               "0.585"  ""       
#>   "copper (median [IQR])"   "69.50 [42.50, 127.25]"     "<0.001" "nonnorm"
#>   "alk.phos (median [IQR])" "1345.00 [941.00, 1992.00]" "0.180"  "nonnorm"
#>   "ast (mean (SD))"         "126.30 (60.27)"            "0.598"  ""       
#>   "trig (median [IQR])"     "113.00 [84.00, 157.00]"    "0.058"  "nonnorm"
#>   "platelet (mean (SD))"    "267.95 (92.20)"            "0.362"  ""       
#>   "protime (median [IQR])"  "10.60 [9.90, 11.25]"       "0.081"  "nonnorm"