Format and print CatTable class objects

print method for the CatTable class objects created by CreateCatTable function.

# S3 method for CatTable
print(
  x,
  digits = 1,
  pDigits = 3,
  quote = FALSE,
  missing = FALSE,
  explain = TRUE,
  printToggle = TRUE,
  noSpaces = FALSE,
  format = c("fp", "f", "p", "pf")[1],
  showAllLevels = FALSE,
  cramVars = NULL,
  dropEqual = FALSE,
  test = TRUE,
  exact = NULL,
  smd = FALSE,
  CrossTable = FALSE,
  formatOptions = list(scientific = FALSE),
  ...
)

Arguments

x
Object returned by CreateCatTable function.

digits
Number of digits to print in the table.

pDigits
Number of digits to print for p-values (also used for standardized mean differences).

quote
Whether to show everything in quotes. The default is FALSE. If TRUE, everything including the row and column names are quoted so that you can copy it to Excel easily.

missing
Whether to show missing data information.

explain
Whether to add explanation to the variable names, i.e., (%) is added to the variable names when percentage is shown.

printToggle
Whether to print the output. If FALSE, no output is created, and a matrix is invisibly returned.

noSpaces
Whether to remove spaces added for alignment. Use this option if you prefer to align numbers yourself in other software.

format
The default is "fp" frequency (percentage). You can also choose from "f" frequency only, "p" percentage only, and "pf" percentage (frequency).

showAllLevels
Whether to show all levels. FALSE by default, i.e., for 2-level categorical variables, only the higher level is shown to avoid redundant information.

cramVars
A character vector to specify the two-level categorical variables, for which both levels should be shown in one row.

dropEqual
Whether to drop " = second level name" description indicating which level is shown for two-level categorical variables.

test
Whether to show p-values. TRUE by default. If FALSE, only the numerical summaries are shown.

exact
A character vector to specify the variables for which the p-values should be those of exact tests. By default all p-values are from large sample approximation tests (chisq.test).

smd
Whether to show standardized mean differences. FALSE by default. If there are more than one contrasts, the average of all possible standardized mean differences is shown. For individual contrasts, use summary.

CrossTable
Whether to show the cross table objects held internally using gmodels::CrossTable function. This will give an output similar to the PROC FREQ in SAS.

formatOptions
A list of options, which will be passed to format. Can be used to modify the big.mark, decimal.mark, big.interval etc. The default is list(scientific = FALSE). The options digits, nsmall, justify and trim are not available. (Experimental)

...
For compatibility with generic. Ignored.

Value

A matrix object containing what you see is also invisibly returned. This can be assinged a name and exported via write.csv.

See also

Author

Kazuki Yoshida, Alexander Bartel

Examples

## Load
library(tableone)

## Load Mayo Clinic Primary Biliary Cirrhosis Data
library(survival)
data(pbc)
## Check variables
head(pbc)

#>   id time status trt      age sex ascites hepato spiders edema bili chol
#> 1  1  400      2   1 58.76523   f       1      1       1   1.0 14.5  261
#> 2  2 4500      0   1 56.44627   f       0      1       1   0.0  1.1  302
#> 3  3 1012      2   1 70.07255   m       0      0       0   0.5  1.4  176
#> 4  4 1925      2   1 54.74059   f       0      1       1   0.5  1.8  244
#> 5  5 1504      1   2 38.10541   f       0      1       1   0.0  3.4  279
#> 6  6 2503      2   2 66.25873   f       0      1       0   0.0  0.8  248
#>   albumin copper alk.phos    ast trig platelet protime stage
#> 1    2.60    156   1718.0 137.95  172      190    12.2     4
#> 2    4.14     54   7394.8 113.52   88      221    10.6     3
#> 3    3.48    210    516.0  96.10   55      151    12.0     4
#> 4    2.54     64   6121.8  60.63   92      183    10.3     4
#> 5    3.53    143    671.0 113.15   72      136    10.9     3
#> 6    3.98     50    944.0  93.00   63       NA    11.0     3

## Create an overall table for categorical variables
catVars <- c("status","ascites","hepato","spiders","edema","stage")
catTableOverall <- CreateCatTable(vars = catVars, data = pbc)

## Simply typing the object name will invoke the print.CatTable method,
## which will show the sample size, frequencies and percentages.
## For 2-level variables, only the higher level is shown for simplicity.
catTableOverall

#>                  
#>                   Overall    
#>   n               418        
#>   status (%)                 
#>      0            232 (55.5) 
#>      1             25 ( 6.0) 
#>      2            161 (38.5) 
#>   ascites = 1 (%)  24 ( 7.7) 
#>   hepato = 1 (%)  160 (51.3) 
#>   spiders = 1 (%)  90 (28.8) 
#>   edema (%)                  
#>      0            354 (84.7) 
#>      0.5           44 (10.5) 
#>      1             20 ( 4.8) 
#>   stage (%)                  
#>      1             21 ( 5.1) 
#>      2             92 (22.3) 
#>      3            155 (37.6) 
#>      4            144 (35.0) 

## If you need to show both levels for some 2-level factors, use cramVars
print(catTableOverall, cramVars = "hepato")

#>                   
#>                    Overall             
#>   n                418                 
#>   status (%)                           
#>      0                 232 (55.5)      
#>      1                  25 ( 6.0)      
#>      2                 161 (38.5)      
#>   ascites = 1 (%)       24 ( 7.7)      
#>   hepato = 0/1 (%) 152/160 (48.7/51.3) 
#>   spiders = 1 (%)       90 (28.8)      
#>   edema (%)                            
#>      0                 354 (84.7)      
#>      0.5                44 (10.5)      
#>      1                  20 ( 4.8)      
#>   stage (%)                            
#>      1                  21 ( 5.1)      
#>      2                  92 (22.3)      
#>      3                 155 (37.6)      
#>      4                 144 (35.0)      

## Use the showAllLevels argument to see all levels for all variables.
print(catTableOverall, showAllLevels = TRUE)

#>              
#>               level Overall    
#>   n                 418        
#>   status (%)  0     232 (55.5) 
#>               1      25 ( 6.0) 
#>               2     161 (38.5) 
#>   ascites (%) 0     288 (92.3) 
#>               1      24 ( 7.7) 
#>   hepato (%)  0     152 (48.7) 
#>               1     160 (51.3) 
#>   spiders (%) 0     222 (71.2) 
#>               1      90 (28.8) 
#>   edema (%)   0     354 (84.7) 
#>               0.5    44 (10.5) 
#>               1      20 ( 4.8) 
#>   stage (%)   1      21 ( 5.1) 
#>               2      92 (22.3) 
#>               3     155 (37.6) 
#>               4     144 (35.0) 

## You can choose form frequencies ("f") and/or percentages ("p") or both.
## "fp" frequency (percentage) is the default. Row names change accordingly.
print(catTableOverall, format = "f")

#>              
#>               Overall
#>   n           418    
#>   status             
#>      0        232    
#>      1         25    
#>      2        161    
#>   ascites = 1  24    
#>   hepato = 1  160    
#>   spiders = 1  90    
#>   edema              
#>      0        354    
#>      0.5       44    
#>      1         20    
#>   stage              
#>      1         21    
#>      2         92    
#>      3        155    
#>      4        144    

print(catTableOverall, format = "p")

#>                  
#>                   Overall
#>   n               418    
#>   status (%)             
#>      0            55.5   
#>      1             6.0   
#>      2            38.5   
#>   ascites = 1 (%)  7.7   
#>   hepato = 1 (%)  51.3   
#>   spiders = 1 (%) 28.8   
#>   edema (%)              
#>      0            84.7   
#>      0.5          10.5   
#>      1             4.8   
#>   stage (%)              
#>      1             5.1   
#>      2            22.3   
#>      3            37.6   
#>      4            35.0   

## To further examine the variables, use the summary.CatTable method,
## which will show more details.
summary(catTableOverall)

#> strata: Overall
#>      var   n miss p.miss level freq percent cum.percent
#>   status 418    0    0.0     0  232    55.5        55.5
#>                              1   25     6.0        61.5
#>                              2  161    38.5       100.0
#>                                                        
#>  ascites 418  106   25.4     0  288    92.3        92.3
#>                              1   24     7.7       100.0
#>                                                        
#>   hepato 418  106   25.4     0  152    48.7        48.7
#>                              1  160    51.3       100.0
#>                                                        
#>  spiders 418  106   25.4     0  222    71.2        71.2
#>                              1   90    28.8       100.0
#>                                                        
#>    edema 418    0    0.0     0  354    84.7        84.7
#>                            0.5   44    10.5        95.2
#>                              1   20     4.8       100.0
#>                                                        
#>    stage 418    6    1.4     1   21     5.1         5.1
#>                              2   92    22.3        27.4
#>                              3  155    37.6        65.0
#>                              4  144    35.0       100.0
#>                                                        

## The table can be stratified by one or more variables
catTableBySexTrt <- CreateCatTable(vars = catVars,
                                   strata = c("sex","trt"), data = pbc)

## print now includes p-values which are by default calculated by chisq.test.
## It is formatted at the decimal place specified by the pDigits argument
## (3 by default). It does <0.001 for you.
catTableBySexTrt

#>                  Stratified by sex:trt
#>                   m:1        f:1         m:2        f:2         p      test
#>   n               21         137         15         139                    
#>   status (%)                                                     0.033     
#>      0             4 (19.0)   79 (57.7)   7 (46.7)   78 (56.1)             
#>      1             3 (14.3)    7 ( 5.1)   0 ( 0.0)    9 ( 6.5)             
#>      2            14 (66.7)   51 (37.2)   8 (53.3)   52 (37.4)             
#>   ascites = 1 (%)  1 ( 4.8)   13 ( 9.5)   2 (13.3)    8 ( 5.8)   0.516     
#>   hepato = 1 (%)  12 (57.1)   61 (44.5)   9 (60.0)   78 (56.1)   0.208     
#>   spiders = 1 (%)  3 (14.3)   42 (30.7)   1 ( 6.7)   44 (31.7)   0.089     
#>   edema (%)                                                      0.906     
#>      0            17 (81.0)  115 (83.9)  12 (80.0)  119 (85.6)             
#>      0.5           3 (14.3)   13 ( 9.5)   1 ( 6.7)   12 ( 8.6)             
#>      1             1 ( 4.8)    9 ( 6.6)   2 (13.3)    8 ( 5.8)             
#>   stage (%)                                                      0.646     
#>      1             2 ( 9.5)   10 ( 7.3)   1 ( 6.7)    3 ( 2.2)             
#>      2             4 (19.0)   31 (22.6)   2 (13.3)   30 (21.6)             
#>      3             7 (33.3)   49 (35.8)   5 (33.3)   59 (42.4)             
#>      4             8 (38.1)   47 (34.3)   7 (46.7)   47 (33.8)             

## The exact argument toggles the p-values to the exact test result from
## fisher.test. It will show which ones are from exact tests.
print(catTableBySexTrt, exact = "ascites")

#>                  Stratified by sex:trt
#>                   m:1        f:1         m:2        f:2         p      test 
#>   n               21         137         15         139                     
#>   status (%)                                                     0.033      
#>      0             4 (19.0)   79 (57.7)   7 (46.7)   78 (56.1)              
#>      1             3 (14.3)    7 ( 5.1)   0 ( 0.0)    9 ( 6.5)              
#>      2            14 (66.7)   51 (37.2)   8 (53.3)   52 (37.4)              
#>   ascites = 1 (%)  1 ( 4.8)   13 ( 9.5)   2 (13.3)    8 ( 5.8)   0.435 exact
#>   hepato = 1 (%)  12 (57.1)   61 (44.5)   9 (60.0)   78 (56.1)   0.208      
#>   spiders = 1 (%)  3 (14.3)   42 (30.7)   1 ( 6.7)   44 (31.7)   0.089      
#>   edema (%)                                                      0.906      
#>      0            17 (81.0)  115 (83.9)  12 (80.0)  119 (85.6)              
#>      0.5           3 (14.3)   13 ( 9.5)   1 ( 6.7)   12 ( 8.6)              
#>      1             1 ( 4.8)    9 ( 6.6)   2 (13.3)    8 ( 5.8)              
#>   stage (%)                                                      0.646      
#>      1             2 ( 9.5)   10 ( 7.3)   1 ( 6.7)    3 ( 2.2)              
#>      2             4 (19.0)   31 (22.6)   2 (13.3)   30 (21.6)              
#>      3             7 (33.3)   49 (35.8)   5 (33.3)   59 (42.4)              
#>      4             8 (38.1)   47 (34.3)   7 (46.7)   47 (33.8)              

## summary now includes both types of p-values
summary(catTableBySexTrt)

#> sex: m
#> trt: 1
#>      var  n miss p.miss level freq percent cum.percent
#>   status 21    0    0.0     0    4    19.0        19.0
#>                             1    3    14.3        33.3
#>                             2   14    66.7       100.0
#>                                                       
#>  ascites 21    0    0.0     0   20    95.2        95.2
#>                             1    1     4.8       100.0
#>                                                       
#>   hepato 21    0    0.0     0    9    42.9        42.9
#>                             1   12    57.1       100.0
#>                                                       
#>  spiders 21    0    0.0     0   18    85.7        85.7
#>                             1    3    14.3       100.0
#>                                                       
#>    edema 21    0    0.0     0   17    81.0        81.0
#>                           0.5    3    14.3        95.2
#>                             1    1     4.8       100.0
#>                                                       
#>    stage 21    0    0.0     1    2     9.5         9.5
#>                             2    4    19.0        28.6
#>                             3    7    33.3        61.9
#>                             4    8    38.1       100.0
#>                                                       
#> ------------------------------------------------------------ 
#> sex: f
#> trt: 1
#>      var   n miss p.miss level freq percent cum.percent
#>   status 137    0    0.0     0   79    57.7        57.7
#>                              1    7     5.1        62.8
#>                              2   51    37.2       100.0
#>                                                        
#>  ascites 137    0    0.0     0  124    90.5        90.5
#>                              1   13     9.5       100.0
#>                                                        
#>   hepato 137    0    0.0     0   76    55.5        55.5
#>                              1   61    44.5       100.0
#>                                                        
#>  spiders 137    0    0.0     0   95    69.3        69.3
#>                              1   42    30.7       100.0
#>                                                        
#>    edema 137    0    0.0     0  115    83.9        83.9
#>                            0.5   13     9.5        93.4
#>                              1    9     6.6       100.0
#>                                                        
#>    stage 137    0    0.0     1   10     7.3         7.3
#>                              2   31    22.6        29.9
#>                              3   49    35.8        65.7
#>                              4   47    34.3       100.0
#>                                                        
#> ------------------------------------------------------------ 
#> sex: m
#> trt: 2
#>      var  n miss p.miss level freq percent cum.percent
#>   status 15    0    0.0     0    7    46.7        46.7
#>                             1    0     0.0        46.7
#>                             2    8    53.3       100.0
#>                                                       
#>  ascites 15    0    0.0     0   13    86.7        86.7
#>                             1    2    13.3       100.0
#>                                                       
#>   hepato 15    0    0.0     0    6    40.0        40.0
#>                             1    9    60.0       100.0
#>                                                       
#>  spiders 15    0    0.0     0   14    93.3        93.3
#>                             1    1     6.7       100.0
#>                                                       
#>    edema 15    0    0.0     0   12    80.0        80.0
#>                           0.5    1     6.7        86.7
#>                             1    2    13.3       100.0
#>                                                       
#>    stage 15    0    0.0     1    1     6.7         6.7
#>                             2    2    13.3        20.0
#>                             3    5    33.3        53.3
#>                             4    7    46.7       100.0
#>                                                       
#> ------------------------------------------------------------ 
#> sex: f
#> trt: 2
#>      var   n miss p.miss level freq percent cum.percent
#>   status 139    0    0.0     0   78    56.1        56.1
#>                              1    9     6.5        62.6
#>                              2   52    37.4       100.0
#>                                                        
#>  ascites 139    0    0.0     0  131    94.2        94.2
#>                              1    8     5.8       100.0
#>                                                        
#>   hepato 139    0    0.0     0   61    43.9        43.9
#>                              1   78    56.1       100.0
#>                                                        
#>  spiders 139    0    0.0     0   95    68.3        68.3
#>                              1   44    31.7       100.0
#>                                                        
#>    edema 139    0    0.0     0  119    85.6        85.6
#>                            0.5   12     8.6        94.2
#>                              1    8     5.8       100.0
#>                                                        
#>    stage 139    0    0.0     1    3     2.2         2.2
#>                              2   30    21.6        23.7
#>                              3   59    42.4        66.2
#>                              4   47    33.8       100.0
#>                                                        
#> 
#> p-values
#>            pApprox     pExact
#> status  0.03269881 0.02420814
#> ascites 0.51569424 0.43540746
#> hepato  0.20805680 0.20480121
#> spiders 0.08898575 0.08569292
#> edema   0.90584197 0.84833223
#> stage   0.64630048         NA
#> 
#> Standardize mean differences
#>           average    1 vs 2     1 vs 3     1 vs 4    2 vs 3     2 vs 4
#> status  0.5770647 0.8748771 0.79644734 0.83097631 0.4375075 0.06043662
#> ascites 0.1755851 0.1845375 0.30219043 0.04452096 0.1211310 0.14105454
#> hepato  0.1598214 0.2544151 0.05802589 0.02073917 0.3135916 0.23336861
#> spiders 0.4018377 0.3999925 0.25073566 0.42200838 0.6471968 0.02154469
#> edema   0.2122338 0.1624118 0.37625250 0.18063879 0.2420277 0.04705137
#> stage   0.2773411 0.1333844 0.21945509 0.35632272 0.3000785 0.26275030
#>             3 vs 4
#> status  0.46214324
#> ascites 0.26007640
#> hepato  0.07878788
#> spiders 0.66954801
#> edema   0.26502087
#> stage   0.39205566

## If your work flow includes copying to Excel and Word when writing manuscripts,
## you may benefit from the quote argument. This will quote everything so that
## Excel does not mess up the cells.
print(catTableBySexTrt, exact = "ascites", quote = TRUE)

#>                    "Stratified by sex:trt"
#>  ""                 "m:1"        "f:1"         "m:2"        "f:2"        
#>   "n"               "21"         "137"         "15"         "139"        
#>   "status (%)"      ""           ""            ""           ""           
#>   "   0"            " 4 (19.0) " " 79 (57.7) " " 7 (46.7) " " 78 (56.1) "
#>   "   1"            " 3 (14.3) " "  7 ( 5.1) " " 0 ( 0.0) " "  9 ( 6.5) "
#>   "   2"            "14 (66.7) " " 51 (37.2) " " 8 (53.3) " " 52 (37.4) "
#>   "ascites = 1 (%)" " 1 ( 4.8) " " 13 ( 9.5) " " 2 (13.3) " "  8 ( 5.8) "
#>   "hepato = 1 (%)"  "12 (57.1) " " 61 (44.5) " " 9 (60.0) " " 78 (56.1) "
#>   "spiders = 1 (%)" " 3 (14.3) " " 42 (30.7) " " 1 ( 6.7) " " 44 (31.7) "
#>   "edema (%)"       ""           ""            ""           ""           
#>   "   0"            "17 (81.0) " "115 (83.9) " "12 (80.0) " "119 (85.6) "
#>   "   0.5"          " 3 (14.3) " " 13 ( 9.5) " " 1 ( 6.7) " " 12 ( 8.6) "
#>   "   1"            " 1 ( 4.8) " "  9 ( 6.6) " " 2 (13.3) " "  8 ( 5.8) "
#>   "stage (%)"       ""           ""            ""           ""           
#>   "   1"            " 2 ( 9.5) " " 10 ( 7.3) " " 1 ( 6.7) " "  3 ( 2.2) "
#>   "   2"            " 4 (19.0) " " 31 (22.6) " " 2 (13.3) " " 30 (21.6) "
#>   "   3"            " 7 (33.3) " " 49 (35.8) " " 5 (33.3) " " 59 (42.4) "
#>   "   4"            " 8 (38.1) " " 47 (34.3) " " 7 (46.7) " " 47 (33.8) "
#>                    "Stratified by sex:trt"
#>  ""                 "p"      "test" 
#>   "n"               ""       ""     
#>   "status (%)"      " 0.033" ""     
#>   "   0"            ""       ""     
#>   "   1"            ""       ""     
#>   "   2"            ""       ""     
#>   "ascites = 1 (%)" " 0.435" "exact"
#>   "hepato = 1 (%)"  " 0.208" ""     
#>   "spiders = 1 (%)" " 0.089" ""     
#>   "edema (%)"       " 0.906" ""     
#>   "   0"            ""       ""     
#>   "   0.5"          ""       ""     
#>   "   1"            ""       ""     
#>   "stage (%)"       " 0.646" ""     
#>   "   1"            ""       ""     
#>   "   2"            ""       ""     
#>   "   3"            ""       ""     
#>   "   4"            ""       ""     

## If you want to center-align values in Word, use noSpaces option.
print(catTableBySexTrt, exact = "ascites", quote = TRUE, noSpaces = TRUE)

#>                    "Stratified by sex:trt"
#>  ""                 "m:1"       "f:1"        "m:2"       "f:2"        "p"    
#>   "n"               "21"        "137"        "15"        "139"        ""     
#>   "status (%)"      ""          ""           ""          ""           "0.033"
#>   "   0"            "4 (19.0)"  "79 (57.7)"  "7 (46.7)"  "78 (56.1)"  ""     
#>   "   1"            "3 (14.3)"  "7 (5.1)"    "0 (0.0)"   "9 (6.5)"    ""     
#>   "   2"            "14 (66.7)" "51 (37.2)"  "8 (53.3)"  "52 (37.4)"  ""     
#>   "ascites = 1 (%)" "1 (4.8)"   "13 (9.5)"   "2 (13.3)"  "8 (5.8)"    "0.435"
#>   "hepato = 1 (%)"  "12 (57.1)" "61 (44.5)"  "9 (60.0)"  "78 (56.1)"  "0.208"
#>   "spiders = 1 (%)" "3 (14.3)"  "42 (30.7)"  "1 (6.7)"   "44 (31.7)"  "0.089"
#>   "edema (%)"       ""          ""           ""          ""           "0.906"
#>   "   0"            "17 (81.0)" "115 (83.9)" "12 (80.0)" "119 (85.6)" ""     
#>   "   0.5"          "3 (14.3)"  "13 (9.5)"   "1 (6.7)"   "12 (8.6)"   ""     
#>   "   1"            "1 (4.8)"   "9 (6.6)"    "2 (13.3)"  "8 (5.8)"    ""     
#>   "stage (%)"       ""          ""           ""          ""           "0.646"
#>   "   1"            "2 (9.5)"   "10 (7.3)"   "1 (6.7)"   "3 (2.2)"    ""     
#>   "   2"            "4 (19.0)"  "31 (22.6)"  "2 (13.3)"  "30 (21.6)"  ""     
#>   "   3"            "7 (33.3)"  "49 (35.8)"  "5 (33.3)"  "59 (42.4)"  ""     
#>   "   4"            "8 (38.1)"  "47 (34.3)"  "7 (46.7)"  "47 (33.8)"  ""     
#>                    "Stratified by sex:trt"
#>  ""                 "test" 
#>   "n"               ""     
#>   "status (%)"      ""     
#>   "   0"            ""     
#>   "   1"            ""     
#>   "   2"            ""     
#>   "ascites = 1 (%)" "exact"
#>   "hepato = 1 (%)"  ""     
#>   "spiders = 1 (%)" ""     
#>   "edema (%)"       ""     
#>   "   0"            ""     
#>   "   0.5"          ""     
#>   "   1"            ""     
#>   "stage (%)"       ""     
#>   "   1"            ""     
#>   "   2"            ""     
#>   "   3"            ""     
#>   "   4"            ""