Format and print svyContTable class objects

print method for the svyContTable class objects created by CreateContTable function.

# S3 method for svyContTable
print(
  x,
  digits = 2,
  pDigits = 3,
  quote = FALSE,
  missing = FALSE,
  explain = TRUE,
  printToggle = TRUE,
  noSpaces = FALSE,
  nonnormal = NULL,
  minMax = FALSE,
  insertLevel = FALSE,
  test = TRUE,
  smd = FALSE,
  formatOptions = list(scientific = FALSE),
  ...
)

Arguments

x
Object returned by svyCreateContTable function.

digits
Number of digits to print in the table.

pDigits
Number of digits to print for p-values (also used for standardized mean differences).

quote
Whether to show everything in quotes. The default is FALSE. If TRUE, everything including the row and column names are quoted so that you can copy it to Excel easily.

missing
Whether to show missing data information.

explain
Whether to add explanation to the variable names, i.e., (mean (SD) or median [IQR]) is added to the variable names.

printToggle
Whether to print the output. If FALSE, no output is created, and a matrix is invisibly returned.

noSpaces
Whether to remove spaces added for alignment. Use this option if you prefer to align numbers yourself in other software.

nonnormal
A character vector to specify the variables for which the p-values should be those of nonparametric tests. By default all p-values are from normal assumption-based tests (oneway.test).

minMax
Whether to use [min,max] instead of [p25,p75] for nonnormal variables. The default is FALSE.

insertLevel
Whether to add an empty level column to the left of strata.

test
Whether to show p-values. TRUE by default. If FALSE, only the numerical summaries are shown.

smd
Whether to show standardized mean differences. FALSE by default. If there are more than one contrasts, the average of all possible standardized mean differences is shown. For individual contrasts, use summary.

formatOptions
A list of options, which will be passed to format. Can be used to modify the big.mark, decimal.mark, big.interval etc. The default is list(scientific = FALSE). The options digits, nsmall, justify and trim are not available. (Experimental)

...
For compatibility with generic. Ignored.

Value

A matrix object containing what you see is also invisibly returned. This can be assinged a name and exported via write.csv.

See also

Examples

## See the examples for svyCreateTableOne()