Create an object summarizing both continuous and categorical variables for weighted data

Create an object summarizing all baseline variables (both continuous and categorical) optionally stratifying by one or more startifying variables and performing statistical tests. The object gives a table that is easy to use in medical research papers.

Usage

svyCreateTableOne(
  vars,
  strata,
  data,
  factorVars,
  includeNA = FALSE,
  test = TRUE,
  testApprox = svyTestChisq,
  argsApprox = NULL,
  testNormal = svyTestNormal,
  argsNormal = list(method = "Wald"),
  testNonNormal = svyTestNonNormal,
  argsNonNormal = NULL,
  smd = TRUE,
  addOverall = FALSE
)

Arguments

vars

Variables to be summarized given as a character vector. Factors are handled as categorical variables, whereas numeric variables are handled as continuous variables. If empty, all variables in the survey design object specified in the data argument are used.

strata

Stratifying (grouping) variable name(s) given as a character vector. If omitted, the overall results are returned.

data

A survey design object in which these variables exist. All variables (both vars and strata) must be in this survey design object. It is created with the svydesign function in the survey package.

factorVars

Numerically coded variables that should be handled as categorical variables given as a character vector. Do not include factors, unless you need to relevel them by removing empty levels. If omitted, only factors are considered categorical variables. The variables specified here must also be specified in the vars argument.

includeNA

If TRUE, NA is handled as a regular factor level rather than missing. NA is shown as the last factor level in the table. Only effective for categorical variables.

test

If TRUE, as in the default and there are more than two groups, groupwise comparisons are performed.

testApprox

A function used to perform the large sample approximation based tests. The default is svychisq.

argsApprox

A named list of arguments passed to the function specified in testApprox.

testNormal

A function used to perform the normal assumption based tests. The default is multiple degrees of freedom test using svyglm and regTermTest. This is equivalent of the svyttest when there are only two groups.

argsNormal

A named list of arguments passed to the function specified in testNormal.

testNonNormal

A function used to perform the nonparametric tests. The default is svyranktest.

argsNonNormal

A named list of arguments passed to the function specified in testNonNormal.

smd

If TRUE, as in the default and there are more than two groups, standardized mean differences for all pairwise comparisons are calculated.

addOverall

(optional, only used if strata are supplied) Adds an overall column to the table. Smd and p-value calculations are performed using only the stratifed clolumns.

Value

An object of class svyTableOne, which is a list of three objects.

ContTable

an object of class svyContTable, containing continuous variables only

CatTable

an object of class svyCatTable, containing categorical variables only

MetaData

list of metadata regarding variables

Details

See the details for CreateTableOne.

See also

Author

Kazuki Yoshida

Examples

## Load packages
library(tableone)
library(survey)
#> Warning: package 'survey' was built under R version 4.1.3
#> Loading required package: grid
#> Loading required package: Matrix
#> 
#> Attaching package: 'survey'
#> The following object is masked from 'package:graphics':
#> 
#>     dotchart

## Create a weighted survey design object
data(nhanes)
nhanesSvy <- svydesign(ids = ~ SDMVPSU, strata = ~ SDMVSTRA, weights = ~ WTMEC2YR,
                       nest = TRUE, data = nhanes)

## Create a table object
## factorVars are converted to factors; no need for variables already factors
## strata will stratify summaries; leave it unspecified for overall summaries
tab1 <- svyCreateTableOne(vars = c("HI_CHOL","race","agecat","RIAGENDR"),
                          strata = "RIAGENDR", data = nhanesSvy,
                          factorVars = c("race","RIAGENDR"))

## Detailed output
summary(tab1)
#> 
#>      ### Summary of continuous variables ###
#> 
#> RIAGENDR: 1
#>             n  miss p.miss mean  sd median p25 p75 min max
#> HI_CHOL 1e+08 1e+07      7  0.1 0.3      0   0   0   0   1
#> ------------------------------------------------------------ 
#> RIAGENDR: 2
#>             n  miss p.miss mean  sd median p25 p75 min max
#> HI_CHOL 1e+08 1e+07      8  0.1 0.3      0   0   0   0   1
#> 
#> p-values
#>             pNormal  pNonNormal
#> HI_CHOL 0.009221135 0.009221135
#> 
#> Standardize mean differences
#>             1 vs 2
#> HI_CHOL 0.07093003
#> 
#> =======================================================================================
#> 
#>      ### Summary of categorical variables ### 
#> 
#> RIAGENDR: 1
#>       var           n miss p.miss    level        freq percent cum.percent
#>      race 134944553.9  0.0    0.0        1  21381884.2    15.8        15.8
#>                                          2  89315751.4    66.2        82.0
#>                                          3  15045455.5    11.1        93.2
#>                                          4   9201462.9     6.8       100.0
#>                                                                           
#>    agecat 134944553.9  0.0    0.0   (0,19]  29299546.1    21.7        21.7
#>                                    (19,39]  40497613.1    30.0        51.7
#>                                    (39,59]  41053579.4    30.4        82.1
#>                                   (59,Inf]  24093815.3    17.9       100.0
#>                                                                           
#>  RIAGENDR 134944553.9  0.0    0.0        1 134944553.9   100.0       100.0
#>                                          2         0.0     0.0       100.0
#>                                                                           
#> ------------------------------------------------------------ 
#> RIAGENDR: 2
#>       var           n miss p.miss    level        freq percent cum.percent
#>      race 141591892.0  0.0    0.0        1  20251367.4    14.3        14.3
#>                                          2  92486945.1    65.3        79.6
#>                                          3  17967228.3    12.7        92.3
#>                                          4  10886351.1     7.7       100.0
#>                                                                           
#>    agecat 141591892.0  0.0    0.0   (0,19]  28150760.5    19.9        19.9
#>                                    (19,39]  40640361.5    28.7        48.6
#>                                    (39,59]  42817044.0    30.2        78.8
#>                                   (59,Inf]  29983725.9    21.2       100.0
#>                                                                           
#>  RIAGENDR 141591892.0  0.0    0.0        1         0.0     0.0         0.0
#>                                          2 141591892.0   100.0       100.0
#>                                                                           
#> 
#> p-values
#>               pApprox pExact
#> race     4.162884e-02     NA
#> agecat   1.175703e-03     NA
#> RIAGENDR 2.688704e-49     NA
#> 
#> Standardize mean differences
#>              1 vs 2
#> race     0.06878329
#> agecat   0.08873652
#> RIAGENDR        NaN

## Default formatted printing
tab1
#>                      Stratified by RIAGENDR
#>                       1                    2                     p      test
#>   n                   134944553.92         141591892.00                     
#>   HI_CHOL (mean (SD))         0.10 (0.30)          0.12 (0.33)    0.009     
#>   race (%)                                                        0.042     
#>      1                  21381884.2 (15.8)    20251367.4 ( 14.3)             
#>      2                  89315751.4 (66.2)    92486945.1 ( 65.3)             
#>      3                  15045455.5 (11.1)    17967228.3 ( 12.7)             
#>      4                   9201462.9 ( 6.8)    10886351.1 (  7.7)             
#>   agecat (%)                                                      0.001     
#>      (0,19]             29299546.1 (21.7)    28150760.5 ( 19.9)             
#>      (19,39]            40497613.1 (30.0)    40640361.5 ( 28.7)             
#>      (39,59]            41053579.4 (30.4)    42817044.0 ( 30.2)             
#>      (59,Inf]           24093815.3 (17.9)    29983725.9 ( 21.2)             
#>   RIAGENDR = 2 (%)             0.0 ( 0.0)   141591892.0 (100.0)  <0.001     

## nonnormal specifies variables to be shown as median [IQR]
print(tab1, nonnormal = "HI_CHOL", contDigits = 3, catDigits = 2,
      pDigits = 4, smd = TRUE)
#>                         Stratified by RIAGENDR
#>                          1                           
#>   n                      134944553.923               
#>   HI_CHOL (median [IQR])         0.000 [0.000, 0.000]
#>   race (%)                                           
#>      1                     21381884.19 (15.84)       
#>      2                     89315751.41 (66.19)       
#>      3                     15045455.46 (11.15)       
#>      4                      9201462.86 ( 6.82)       
#>   agecat (%)                                         
#>      (0,19]                29299546.11 (21.71)       
#>      (19,39]               40497613.07 (30.01)       
#>      (39,59]               41053579.41 (30.42)       
#>      (59,Inf]              24093815.33 (17.85)       
#>   RIAGENDR = 2 (%)                0.00 ( 0.00)       
#>                         Stratified by RIAGENDR
#>                          2                            p       test    SMD    
#>   n                      141591891.998                                       
#>   HI_CHOL (median [IQR])         0.000 [0.000, 0.000]  0.0092 nonnorm  0.0709
#>   race (%)                                             0.0416          0.0688
#>      1                     20251367.39 ( 14.30)                              
#>      2                     92486945.14 ( 65.32)                              
#>      3                     17967228.32 ( 12.69)                              
#>      4                     10886351.15 (  7.69)                              
#>   agecat (%)                                           0.0012          0.0887
#>      (0,19]                28150760.54 ( 19.88)                              
#>      (19,39]               40640361.53 ( 28.70)                              
#>      (39,59]               42817044.01 ( 30.24)                              
#>      (59,Inf]              29983725.90 ( 21.18)                              
#>   RIAGENDR = 2 (%)        141591892.00 (100.00)       <0.0001             NaN

## minMax changes it to median [min, max]
print(tab1, nonnormal = "HI_CHOL", minMax = TRUE, contDigits = 3,
      catDigits = 2, pDigits = 4, smd = TRUE)
#>                           Stratified by RIAGENDR
#>                            1                           
#>   n                        134944553.923               
#>   HI_CHOL (median [range])         0.000 [0.000, 1.000]
#>   race (%)                                             
#>      1                       21381884.19 (15.84)       
#>      2                       89315751.41 (66.19)       
#>      3                       15045455.46 (11.15)       
#>      4                        9201462.86 ( 6.82)       
#>   agecat (%)                                           
#>      (0,19]                  29299546.11 (21.71)       
#>      (19,39]                 40497613.07 (30.01)       
#>      (39,59]                 41053579.41 (30.42)       
#>      (59,Inf]                24093815.33 (17.85)       
#>   RIAGENDR = 2 (%)                  0.00 ( 0.00)       
#>                           Stratified by RIAGENDR
#>                            2                            p       test    SMD    
#>   n                        141591891.998                                       
#>   HI_CHOL (median [range])         0.000 [0.000, 1.000]  0.0092 nonnorm  0.0709
#>   race (%)                                               0.0416          0.0688
#>      1                       20251367.39 ( 14.30)                              
#>      2                       92486945.14 ( 65.32)                              
#>      3                       17967228.32 ( 12.69)                              
#>      4                       10886351.15 (  7.69)                              
#>   agecat (%)                                             0.0012          0.0887
#>      (0,19]                  28150760.54 ( 19.88)                              
#>      (19,39]                 40640361.53 ( 28.70)                              
#>      (39,59]                 42817044.01 ( 30.24)                              
#>      (59,Inf]                29983725.90 ( 21.18)                              
#>   RIAGENDR = 2 (%)          141591892.00 (100.00)       <0.0001             NaN

## showAllLevels can be used tow show levels for all categorical variables
print(tab1, showAllLevels = TRUE, smd = TRUE)
#>                      Stratified by RIAGENDR
#>                       level    1                     2                    
#>   n                            134944553.92          141591892.00         
#>   HI_CHOL (mean (SD))                  0.10 (0.30)           0.12 (0.33)  
#>   race (%)            1          21381884.2 ( 15.8)    20251367.4 ( 14.3) 
#>                       2          89315751.4 ( 66.2)    92486945.1 ( 65.3) 
#>                       3          15045455.5 ( 11.1)    17967228.3 ( 12.7) 
#>                       4           9201462.9 (  6.8)    10886351.1 (  7.7) 
#>   agecat (%)          (0,19]     29299546.1 ( 21.7)    28150760.5 ( 19.9) 
#>                       (19,39]    40497613.1 ( 30.0)    40640361.5 ( 28.7) 
#>                       (39,59]    41053579.4 ( 30.4)    42817044.0 ( 30.2) 
#>                       (59,Inf]   24093815.3 ( 17.9)    29983725.9 ( 21.2) 
#>   RIAGENDR (%)        1         134944553.9 (100.0)           0.0 (  0.0) 
#>                       2                 0.0 (  0.0)   141591892.0 (100.0) 
#>                      Stratified by RIAGENDR
#>                       p      test SMD   
#>   n                                     
#>   HI_CHOL (mean (SD))  0.009       0.071
#>   race (%)             0.042       0.069
#>                                         
#>                                         
#>                                         
#>   agecat (%)           0.001       0.089
#>                                         
#>                                         
#>                                         
#>   RIAGENDR (%)        <0.001         NaN
#>                                         

## To see all printing options
?print.TableOne

## To examine categorical variables only
tab1$CatTable
#>                   Stratified by RIAGENDR
#>                    1                  2                    p      test
#>   n                134944553.9        141591892.0                     
#>   race (%)                                                  0.042     
#>      1             21381884.2 (15.8)   20251367.4 ( 14.3)             
#>      2             89315751.4 (66.2)   92486945.1 ( 65.3)             
#>      3             15045455.5 (11.1)   17967228.3 ( 12.7)             
#>      4              9201462.9 ( 6.8)   10886351.1 (  7.7)             
#>   agecat (%)                                                0.001     
#>      (0,19]        29299546.1 (21.7)   28150760.5 ( 19.9)             
#>      (19,39]       40497613.1 (30.0)   40640361.5 ( 28.7)             
#>      (39,59]       41053579.4 (30.4)   42817044.0 ( 30.2)             
#>      (59,Inf]      24093815.3 (17.9)   29983725.9 ( 21.2)             
#>   RIAGENDR = 2 (%)        0.0 ( 0.0)  141591892.0 (100.0)  <0.001     

## To examine continuous variables only
tab1$ContTable
#>                      Stratified by RIAGENDR
#>                       1            2            p      test
#>   n                   134944553.92 141591892.00            
#>   HI_CHOL (mean (SD)) 0.10 (0.30)  0.12 (0.33)   0.009     

## If SMDs are needed as numericals, use ExtractSmd()
ExtractSmd(tab1)
#>              1 vs 2
#> HI_CHOL  0.07093003
#> race     0.06878329
#> agecat   0.08873652
#> RIAGENDR        NaN