Create "Table 1" to describe baseline characteristics

Creates "Table 1", i.e., description of baseline patient characteristics, which is essential in every medical research. Supports both continuous and categorical variables, as well as p-values and standardized mean differences. Weighted data are supported via the survey package. See github for a screencast. tableone was inspired by descriptive statistics functions in Deducer , a Java-based GUI package by Ian Fellows. This package does not require GUI or Java, and intended for command-line users. Most important functions are CreateTableOne and svyCreateTableOne.

Note

Acknowledgement:

Ian Fellows for developing the deducer package, which this package is based on.

Hadley Wickham for packaging advice and for creating tools this package was made with (roxygen2, devtools, testthat).

Yoshinobu Kanda for design advice and for integration into RcmdrPlugin.EZR.

H.Tachibana and Hiroki Matsui for inputs regarding standardized mean differences.

jomuller, Raja Sriswan Mamidi, Atsushi Shiraishi, and Jacques Ropers for bug reports and/or feature suggestions.

Members of the Facebook Organization of R Users for Medical Statistics in Japan (FORUMS-J) for testing pre-release versions and suggestions.

Developmental repository is on github. Your contributions are appreciated.

https://github.com/kaz-yos/tableone

See also

Author

Kazuki Yoshida, Alexander Bartel, Jonathan J Chipman, Justin Bohn, Lucy DAgostino McGowan, Malcom Barrett, Rune Haubo B Christensen, gbouzill

Maintainer: Kazuki Yoshida <kazukiyoshida@mail.harvard.edu>

Examples

## See examples for CreateTableOne and svyCreateTableOne